### 2017-05-22 #266 @whedon
[PRE REVIEW]: Bézier
**Submitting author:** @dhermes (<a href="http://orcid.org/0000-0001-7366-173X">Daniel Hermes</a>)
**Repository:** <a href="https://github.com/dhermes/bezier" target ="_blank">https://github.com/dhermes/bezier</a>
**Version:** 0.3.0
