### 2017-05-22 ad26e6a
Started change over to nox (from tox).
Also dropped 3.4 for 3.6 in unit test runs on CircleCI.
