+++ tox.ini
@@ -1,15 +1,13 @@
 [tox]
-envlist = py27,py34,py35,cover
+envlist = py27,py34,py35
 
 [testenv]
 deps =
-    cover: coverage
     mock >= 1.3.0
     pypy,functional-pypy: git+https://bitbucket.org/pypy/numpy.git
     numpy
     pytest
-    cover: pytest-cov
-    py27,py34,py35,cover: scipy
+    py27,py34,py35: scipy
 commands =
     py.test {posargs} {toxinidir}/tests
 passenv =
@@ -21,31 +19,11 @@ setenv =
     doctest: NO_IMAGES=True
     docs-images: MATPLOTLIBRC={toxinidir}/docs
 
-[pytest]
-addopts = --tb=native
-
-[testenv:cover]
-basepython =
-    python2.7
-commands =
-    py.test \
-      --cov=bezier \
-      --cov=tests \
-      {posargs} \
-      {toxinidir}/tests \
-      {toxinidir}/functional_tests/test_segment_box.py
-
-[testenv:functional]
-basepython =
-    python2.7
-commands =
-    py.test {posargs} {toxinidir}/functional_tests
-
 [testenv:functional-pypy]
 basepython =
     pypy
 commands =
-    {[testenv:functional]commands}
+    py.test {posargs} {toxinidir}/functional_tests
 
 [testenv:docs]
 basepython =
