### 2017-05-22 e5473cf
Downgrading to 2.7.12 on CircleCI.
