### 2017-05-22 f71afb2
Moving docs and doctest build envs to nox.
Somewhat hacky due to https://github.com/jonparrott/nox/issues/28
