+++ tox.ini
@@ -16,7 +16,6 @@ passenv =
 setenv =
     pypy,functional-pypy: MATPLOTLIBRC={toxinidir}/tests
     lint: PYTHONPATH={toxinidir}/functional_tests
-    doctest: NO_IMAGES=True
     docs-images: MATPLOTLIBRC={toxinidir}/docs
 
 [testenv:functional-pypy]
@@ -25,17 +24,9 @@ basepython =
 commands =
     py.test {posargs} {toxinidir}/functional_tests
 
-[testenv:docs]
+[testenv:docs-images]
 basepython =
     python3.5
-commands =
-    {toxinidir}/scripts/build_docs.sh
-deps =
-    --requirement={toxinidir}/docs/requirements.txt
-
-[testenv:doctest]
-basepython =
-    {[testenv:docs]basepython}
 commands =
     sphinx-build -W \
       -b doctest \
@@ -44,15 +35,7 @@ commands =
       {toxinidir}/docs/build/doctest \
       {posargs}
 deps =
-    {[testenv:docs]deps}
-
-[testenv:docs-images]
-basepython =
-    {[testenv:docs]basepython}
-commands =
-    {[testenv:doctest]commands}
-deps =
-    {[testenv:docs]deps}
+    --requirement={toxinidir}/docs/requirements.txt
     matplotlib >= 2.0.0
     seaborn
 
