### 2017-05-22 6ce0034
Moving lint build env. over to nox.
