+++ tox.ini
@@ -15,7 +15,6 @@ passenv =
     docs-images: DISPLAY
 setenv =
     pypy,functional-pypy: MATPLOTLIBRC={toxinidir}/tests
-    lint: PYTHONPATH={toxinidir}/functional_tests
     docs-images: MATPLOTLIBRC={toxinidir}/docs
 
 [testenv:functional-pypy]
@@ -38,35 +37,3 @@ deps =
     --requirement={toxinidir}/docs/requirements.txt
     matplotlib >= 2.0.0
     seaborn
-
-[testenv:lint]
-basepython =
-    python3.5
-commands =
-    python {toxinidir}/scripts/check_readme.py
-    python {toxinidir}/setup.py check \
-      --metadata --restructuredtext --strict
-    flake8 \
-      --import-order-style=google \
-      --application-import-names="bezier,tests" \
-      {toxinidir}/src/bezier {toxinidir}/tests
-    pylint --rcfile pylintrc \
-      --max-module-lines=2463 \
-      {toxinidir}/src/bezier
-    pylint --rcfile pylintrc \
-      --disable=missing-docstring \
-      --disable=protected-access \
-      --disable=too-many-public-methods \
-      --max-module-lines=2324 \
-      {toxinidir}/functional_tests \
-      {toxinidir}/tests
-deps =
-    docutils
-    flake8
-    flake8-import-order
-    matplotlib
-    mock >= 1.3.0
-    numpy
-    Pygments
-    pylint
-    pytest
