### 2017-05-23 3000b5a
Moving docs-images build env. to nox.
Also re-running it with the latest version of matplotlib,
which changed a lot of images.

This also revealed a bug in _plot_helpers.add_patch().
