+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27,py34,py35
+envlist = py27,py35
 
 [testenv]
 deps =
@@ -12,28 +12,11 @@ commands =
     py.test {posargs} {toxinidir}/tests
 passenv =
     MATPLOTLIBRC
-    docs-images: DISPLAY
 setenv =
     pypy,functional-pypy: MATPLOTLIBRC={toxinidir}/tests
-    docs-images: MATPLOTLIBRC={toxinidir}/docs
 
 [testenv:functional-pypy]
 basepython =
     pypy
 commands =
     py.test {posargs} {toxinidir}/functional_tests
-
-[testenv:docs-images]
-basepython =
-    python3.5
-commands =
-    sphinx-build -W \
-      -b doctest \
-      -d {toxinidir}/docs/build/doctrees \
-      {toxinidir}/docs \
-      {toxinidir}/docs/build/doctest \
-      {posargs}
-deps =
-    --requirement={toxinidir}/docs/requirements.txt
-    matplotlib >= 2.0.0
-    seaborn
