### 2017-05-23 ca05be2
Convert final tox envs into nox.
Also

- removing all references to tox.
- fixing "and and" typo in `paper.md`
- making enum34 only required on Python 2.7
