+++ tox.ini
@@ -1,22 +1,0 @@
-[tox]
-envlist = py27,py35
-
-[testenv]
-deps =
-    mock >= 1.3.0
-    pypy,functional-pypy: git+https://bitbucket.org/pypy/numpy.git
-    numpy
-    pytest
-    py27,py34,py35: scipy
-commands =
-    py.test {posargs} {toxinidir}/tests
-passenv =
-    MATPLOTLIBRC
-setenv =
-    pypy,functional-pypy: MATPLOTLIBRC={toxinidir}/tests
-
-[testenv:functional-pypy]
-basepython =
-    pypy
-commands =
-    py.test {posargs} {toxinidir}/functional_tests
