### 2017-05-23 7a86d90
Fixing lint issue in plot helpers unit test.
