### 2017-05-24 #267 @tacaswell
- There are 145 commits to master from the 0.3.0 tag, changes include dependency changes
 - missing link to docs on the github page (ended up using google to find them)
 - could the `DEVELOPMENT.rst` page be included in the html docs?
 - a bit more narrative as to the applications of this package on the main page would be good

Can not get the functional tests to run following the directions (more-or-less with python36) which I assume is due to changes with absolute import

```
23:03 $ python -m pytest functional_tests/
======================================================================================================== test session starts =========================================================================================================
rootdir: /home/tcaswell/src/o/bezier, inifile: tox.ini
plugins: cov-2.4.0, xdist-1.15.0
collected 0 items / 4 errors

=============================================================================================================== ERRORS ===============================================================================================================
_______________________________________________________________________________________ ERROR collecting functional_tests/test_curve_curve.py ________________________________________________________________________________________
ImportError while importing test module '/home/tcaswell/src/o/bezier/functional_tests/test_curve_curve.py'.
Traceback:
functional_tests/test_curve_curve.py:20: in <module>
    import runtime_utils
E   ModuleNotFoundError: No module named 'runtime_utils'
_______________________________________________________________________________________ ERROR collecting functional_tests/test_segment_box.py ________________________________________________________________________________________
ImportError while importing test module '/home/tcaswell/src/o/bezier/functional_tests/test_segment_box.py'.
Traceback:
functional_tests/test_segment_box.py:16: in <module>
    import runtime_utils
E   ModuleNotFoundError: No module named 'runtime_utils'
______________________________________________________________________________________ ERROR collecting functional_tests/test_surface_locate.py ______________________________________________________________________________________
ImportError while importing test module '/home/tcaswell/src/o/bezier/functional_tests/test_surface_locate.py'.
Traceback:
functional_tests/test_surface_locate.py:17: in <module>
    import runtime_utils
E   ModuleNotFoundError: No module named 'runtime_utils'
_____________________________________________________________________________________ ERROR collecting functional_tests/test_surface_surface.py ______________________________________________________________________________________
ImportError while importing test module '/home/tcaswell/src/o/bezier/functional_tests/test_surface_surface.py'.
Traceback:
functional_tests/test_surface_surface.py:21: in <module>
    import runtime_utils
E   ModuleNotFoundError: No module named 'runtime_utils'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 4 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
====================================================================================================== 4 error in 0.83 seconds =======================================================================================================
(dd36) ✘-2 ~/src/o/bezier [0.3.0|✚ 3]

```

but by changing to functional tests and running ` python -m pytest` got them to pass

Getting one failure on the unit tests
```
(dd36) ✘-1 ~/src/o/bezier [0.3.0|✚ 3]
23:17 $ python -m pytest tests/
======================================================================================================== test session starts =========================================================================================================
rootdir: /home/tcaswell/src/o/bezier, inifile: tox.ini
plugins: cov-2.4.0, xdist-1.15.0
collected 345 items

tests/test__base.py .......
tests/test__curve_helpers.py ........................
tests/test__helpers.py .......................
tests/test__intersection_helpers.py ..................................................................
tests/test__plot_helpers.py ..
tests/test__surface_helpers.py .........................................................................................................
tests/test_curve.py .......................................
tests/test_curved_polygon.py ..........
tests/test_surface.py .F...................................................................

============================================================================================================== FAILURES ==============================================================================================================
______________________________________________________________________________________________ TestSurface.test___repr__custom_triangle ______________________________________________________________________________________________
Traceback (most recent call last):
  File "/home/tcaswell/src/o/bezier/tests/test_surface.py", line 124, in test___repr__custom_triangle
    nodes = np.zeros((num_nodes, dimension))
TypeError: 'float' object cannot be interpreted as an integer
================================================================================================ 1 failed, 344 passed in 4.10 seconds ================================================================================================
```
Only got part way through this review due to reading about Barycentric coordinate system and learning some new (to me) math!
