### 2017-05-24 dhermes/bezier#25 @dhermes
Fix functional test build / test env for Python 3
Ref: https://github.com/openjournals/joss-reviews/issues/267
