### 2017-05-24 #267 @dhermes
> There are 145 commits to master from the 0.3.0 tag, changes include dependency changes

Indeed. I have had issues with the Windows build since adding the Fortran speedups. I'd love to do a release but don't want to leave out Windows users.

> missing link to docs on the github page (ended up using google to find them)

There is a badge "docs | latest" in the README. I'm open to suggestions on adding something else / where you'd like to see it added.

![screenshot from 2017-05-23 22-48-11](https://cloud.githubusercontent.com/assets/520669/26388564/eeb9f390-4009-11e7-8f4d-65b365da16be.png)

> could the DEVELOPMENT.rst page be included in the html docs?

Definitely.

> a bit more narrative as to the applications of this package on the main page would be good

- By main page, do you mean http://bezier.readthedocs.io/en/latest/ or the GitHub README?
- When you say "more narrative", do you mean "work with Bezier curves / surfaces / etc." or "here's what a Bezier curve / surface / etc. is and you might want one because <blah>"?

> Can not get the functional tests to run following the directions (more-or-less with python36) which I assume is due to changes with absolute import

Yes I noticed that yesterday when I was updated the CI / testing environment. [Right now][1] they are running in Python 2.7 on CI, but I've filed [an issue][2] to fix and will get to it ASAP tomorrow.

> Getting one failure on the unit tests

I can't reproduce locally or [in CI][3]. Can we discuss your environment a bit more?

> Only got part way through this review due to reading about Barycentric coordinate system and learning some new (to me) math!

They can be really great! One particularly awesome feature of Bezier curves / the Bernstein basis is the convexity property. It allows for lots of nice optimizations. I'm happy to chat about the math if you'd like me to shed some light on things. (The bibliography is a bit long.) I'm certainly not an authority on the subject but there is plenty I do know.

[1]: https://github.com/dhermes/bezier/blob/7a86d904b2e9b48decbd5d6b3fd2340d492b3947/circle.yml#L20
[2]: https://github.com/dhermes/bezier/issues/25
[3]: https://circleci.com/gh/dhermes/bezier
