### 2017-05-24 #267 @tacaswell
RE: link to docs:  :sheep: I was looking for a blue link that said 'documentation' and totally skipped over the badges!

RE: narrative:  both the README and the RTD docs (but content should be very similar).  On the "here's what a Bezier curve / surface / etc. is and you might want one because ..." to address the statement of need.

----

The test failure is in `def test___repr__custom_triangle(self):` and is raised from the the test code due to

```
(dd36) ✘-1 ~/src/o/bezier [0.3.0|✚ 3]
08:48 $ ipython
Python 3.6.1 |Continuum Analytics, Inc.| (default, Mar 22 2017, 19:54:23)
Type "copyright", "credits" or "license" for more information.

IPython 5.2.2 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [1]: import numpy

In [2]: import numpy as np

In [3]: np.zeros((13.0, 7))
<ipython-input-3-f0a5700cdff1> in <module>()
----> 1 np.zeros((13.0, 7))

TypeError: 'float' object cannot be interpreted as an integer

In [4]: np.__version__
Out[4]: '1.12.1'
```
