### 2017-05-24 #267 @dhermes
@tacaswell You're seeing this failure at `HEAD` or at the `0.3.0` tag?
