### 2017-05-24 2e01c54
Updating DEVELOPMENT doc with notes of needed environment variables.
