+++ functional_tests/test_surface_locate.py
@@ -10,6 +10,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from __future__ import absolute_import
+
 try:
     import matplotlib.pyplot as plt
 except ImportError:
