+++ functional_tests/test_implicitized_intersect.py
@@ -10,6 +10,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from __future__ import absolute_import
+
 import numpy as np
 import pytest
 import six
