### 2017-05-24 bc8de58
Updating CircleCI config to include fnl. tests in path.
