### 2017-05-24 d76f6ef
Manually skipping Fortran speedups on Windows in setup.py.
This is so that a 0.4.0 release can go out with Fortran speedups
without blocking on me figuring out Windows builds.
