+++ circle.yml
@@ -27,6 +27,6 @@ deployment:
     tag: /.*/
     owner: dhermes
     commands:
-      - python setup.py sdist bdist_wheel
+      - python setup.py sdist
       - pip install --upgrade twine
       - twine upload dist/*
