### 2017-05-24 3b277fc
Pass Appveyor build.
See: https://www.appveyor.com/docs/build-environment/#miniconda
i.e. `Miniconda36` and `Miniconda36-x64` are valid.

Needed to add `PYTHONPATH=functional_tests/` environment variable
to get imports working correctly in the `cover` and `functional`
environments

Also adding `cov-fail-under` to the coverage script. This is because lines
that import or test Fortran extensions don't get visited (since for now we
don't build the Fortran speedups on Windows).

Also removing

- `conda install m2w64-gcc-fortran`
- `python setup.py build_ext --compiler=msvc --inplace`

for now sinc
[…]
