### 2017-05-24 e08fb94
Updating instructions on creating images from fnl. tests.
