### 2017-05-24 16d3845
Making a few notes about / in `test_implicitized_intersect.py`.
The note was no longer true as of a65f827e2744953fb93cc9b60d7595b43ec57c75,
so it has been removed.

Also re-factoring `test_implicitized_intersect.py` a bit so that
the code does less in each function.
