### 2017-05-24 8615d3e
Adding note about AppVeyor in development doc.
