### 2017-05-24 4675a03
Fixing up indent in bash continuation.
(This is a really useless commit, but I noticed one space where
I meant to have two spaces.)
