### 2017-05-25 02c8e5b
Adding fnl. test images into docs_images.
Also updated the DEVELOPMENT doc to reflect this change and made
a few other minor tweaks (including RAW links to GitHub images, making
a code-block section "rest" == "restructered text", changing a link
and adding / moving some periods).
