### 2017-05-25 8915b04
Adding DEVELOPMENT doc to Sphinx docs.
Also adding "align: center" to images in the DEVELOPMENT doc.
