+++ docs/development.rst
@@ -1,0 +1,1 @@
+../DEVELOPMENT.rst
