### 2017-05-25 #267 @dhermes
FYI @tacaswell I [added][1] the DEVELOPMENT doc to the HTML / Sphinx docs: http://bezier.readthedocs.io/en/latest/development.html

[1]: https://github.com/dhermes/bezier/commit/8915b040cecb62758b3d7056a1847374d0aa4995
