### 2017-05-25 3b3decf
Allowing more generic `_speedup*.so` filename.
Fixes #28.
