### 2017-05-25 dhermes/bezier#25 @dhermes
Fixed (sort of) in

A more "comprehensive" fix would be to catch an `ImportError` when trying to import the modules in the `functional_tests/` directory. However, it would be a lot of repeating code across `test_*.py` modules to wrap for every set of imports.
