### 2017-05-25 2b80757
Upgrade to `nox >= 0.16.1`.
This is so environment variables can be inserted by `nox`
rather than relying on users to set them (this was
intended by `nox` but the feature had a bug).
