+++ circle.yml
@@ -5,7 +5,7 @@ machine:
 dependencies:
   override:
     # Needed for testing (NumPy distutils needed for install)
-    - pip install --upgrade pip numpy nox-automation
+    - pip install --upgrade pip numpy "nox-automation >= 0.16.1"
     # latest as of May 22, 2017: 2.7.13, 3.5.3, 3.6.1
     - pyenv local 2.7.12 3.5.3 3.6.1
 
@@ -13,11 +13,11 @@ test:
   override:
     - nox -s "unit_tests(python_version='3.5')"
     - nox -s "unit_tests(python_version='3.6')"
-    - PYTHONPATH=functional_tests/ nox -s cover
+    - nox -s cover
     - nox -s docs
-    - NO_IMAGES=True nox -s doctest
-    - PYTHONPATH=functional_tests/ nox -s lint
-    - PYTHONPATH=functional_tests/ nox -s "functional(python_version='2.7')"
+    - nox -s doctest
+    - nox -s lint
+    - nox -s "functional(python_version='2.7')"
   post:
     - .nox/cover/bin/pip install coveralls
     - .nox/cover/bin/coveralls
