### 2017-06-01 d090861
Adding LICENSE file to MANIFEST.in.
