+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include README.rst.template
+include README.rst.template LICENSE
 include .f2py_f2cmap
 graft src/bezier
 global-exclude *.pyc
