### 2017-06-06 32d6add
Adding note in setup.py about where else the version need be changed.
