### 2017-06-07 02b94d4
Bringing README and index.rst closer to having the same content.
