### 2017-06-07 a69b7f6
Fixing RST typo (bad link).
