### 2017-06-07 bebc478
Combining usage/getting started sections from index.rst/README.rst.
