+++ docs/index.rst
@@ -41,11 +41,62 @@ code from `GitHub`_ and install from source.
 .. _pip: https://pip.pypa.io
 .. _GitHub: https://github.com/dhermes/bezier/
 
-Usage
------
+Getting Started
+---------------
 
-The B |eacute| zier :doc:`Package <reference/bezier>` documentation provides
-API-level documentation.
+For example, to create a curve:
+
+.. code-block:: python
+
+   >>> nodes1 = np.asfortranarray([
+   ...     [0.0, 0.0],
+   ...     [0.5, 1.0],
+   ...     [1.0, 0.0],
+   ... ])
+   >>> curve1 = bezier.Curve(nodes1, degree=2)
+
+The intersection (points) between two curves can
+also be determined:
+
+.. code-block:: python
+
+   >>> nodes2 = np.asfortranarray([
+   ...     [0.0 ,  0.0],
+   ...     [0.25,  2.0],
+   ...     [0.5 , -2.0],
+   ...     [0.75,  2.0],
+   ...     [1.0 ,  0.0],
+   ... ])
+   >>> curve2 = bezier.Curve.from_nodes(nodes2)
+   >>> intersections = curve1.intersect(curve2)
+   >>> intersections
+   array([[ 0.311...,  0.428...],
+          [ 0.688...,  0.428...],
+          [ 0.      ,  0.      ],
+          [ 1.      ,  0.      ]])
+
+and then we can plot these curves (along with their
+intersections):
+
+.. code-block:: python
+
+   >>> import matplotlib.pyplot as plt
+   >>> import seaborn
+   >>>
+   >>> ax = curve1.plot(num_pts=256)
+   >>> curve2.plot(num_pts=256, ax=ax)
+   >>> ax.plot(intersections[:, 0], intersections[:, 1],
+   ...         marker='o', linestyle='None', color='black')
+   >>> ax.axis('scaled')
+   >>> ax.set_xlim(-0.125, 1.125)
+   >>> ax.set_ylim(-0.0625, 0.625)
+   >>> plt.show()
+
+.. image:: images/test_curves1_and_13.png
+   :align: center
+
+For API-level documentation, check out the B |eacute| zier
+:doc:`Package <reference/bezier>` documentation.
 
 Development
 -----------
