### 2017-06-07 aa4b794
Switching Python code-blocks to doctests in index.rst.
Also

- Changing intersphinx links from Python 3.5 to 3.6
- Updating the "plot" snippets in README since things like ax.plot()
  have return values but we don't want them showing up in the output
