+++ README.rst
@@ -65,10 +65,10 @@ also be determined:
    >>> curve2 = bezier.Curve.from_nodes(nodes2)
    >>> intersections = curve1.intersect(curve2)
    >>> intersections
-   array([[ 0.311...,  0.428...],
-          [ 0.688...,  0.428...],
-          [ 0.      ,  0.      ],
-          [ 1.      ,  0.      ]])
+   array([[ 0.31101776, 0.42857143],
+          [ 0.68898224, 0.42857143],
+          [ 0.        , 0.        ],
+          [ 1.        , 0.        ]])
 
 and then we can plot these curves (along with their
 intersections):
@@ -79,12 +79,13 @@ intersections):
    >>> import seaborn
    >>>
    >>> ax = curve1.plot(num_pts=256)
-   >>> curve2.plot(num_pts=256, ax=ax)
-   >>> ax.plot(intersections[:, 0], intersections[:, 1],
-   ...         marker='o', linestyle='None', color='black')
-   >>> ax.axis('scaled')
-   >>> ax.set_xlim(-0.125, 1.125)
-   >>> ax.set_ylim(-0.0625, 0.625)
+   >>> _ = curve2.plot(num_pts=256, ax=ax)
+   >>> lines = ax.plot(
+   ...     intersections[:, 0], intersections[:, 1],
+   ...     marker='o', linestyle='None', color='black')
+   >>> _ = ax.axis('scaled')
+   >>> _ = ax.set_xlim(-0.125, 1.125)
+   >>> _ = ax.set_ylim(-0.0625, 0.625)
    >>> plt.show()
 
 .. image:: https://cdn.rawgit.com/dhermes/bezier/master/docs/images/test_curves1_and_13.png
