### 2017-06-07 c7e0ca8
Replacing many http links with https equivalents.
