### 2017-06-07 7d16204
Adding a period to docs sentence about LICENSE.
