+++ README.rst
@@ -109,7 +109,7 @@ License
 -------
 
 ``bezier`` is made available under the Apache 2.0 License. For more
-details, see `the LICENSE`_
+details, see `the LICENSE`_.
 
 .. _the LICENSE: https://github.com/dhermes/bezier/blob/master/LICENSE
 
