### 2017-06-07 679d78b
Adding "mock" as a dependency on AppVeyor for `doctest` env.
