### 2017-06-07 057b106
Adding a "Why Bezier?" section to docs.
