### 2017-06-07 c8dc90e
Removing some asserts from main page doctests.
