### 2017-06-07 e11ef0a
Adding "Why Bezier?" section to README as well.
