### 2017-06-07 6d13de8
Docs typo fix: defined -> define.
