### 2017-06-07 9e33f98
Cutting release 0.4.0.
In the process, also removing badges from RTD docs (since they
are only meant for the "latest" build).
