### 2017-06-07 99362f6
Restore badges on RTD main page.
(These are just meant to be removed from "tagged" docs,
i.e. docs for releases. I should definitely get a "better"
way to do this, but this hack will do for now.)
