### 2017-06-07 1a4dc2a
Using 3.6 instead of 2.7 to run functional tests on CircleCI.
