### 2017-06-07 #267 @dhermes
@tacaswell I have added a [Why Bezier?][1] section and also done a `0.4.0` release.

AFAICT I have addressed all of the current feedback, LMK what I need do from here (e.g. if the `0.4.0` release has thrown a wrench into this process).

[1]: https://bezier.readthedocs.io/en/0.4.0/#why-b-eacute-zier
