### 2017-06-08 3afc6e5
Adding verification step to manylinux script.
