+++ scripts/manylinux/build-wheels.sh
@@ -10,24 +10,25 @@ PKG_NAME="bezier";
 # Install a system package required by our library
 yum install -y gcc-gfortran
 
-# Compile wheels
+VERSION_WHITELIST="";
 for PYBIN in /opt/python/*/bin; do
     # H/T: https://stackoverflow.com/a/229606/1068170
-    if [[ "${PYBIN}" == *"26"* ]]; then
-        echo "Python 2.6 (via ${PYBIN}) unsupported";
-        echo "=====================================";
+    if [[ "${PYBIN}" == *"27"* ]]; then
+        VERSION_WHITELIST="${VERSION_WHITELIST} ${PYBIN}";
+    elif [[ "${PYBIN}" == *"35"* ]]; then
+        VERSION_WHITELIST="${VERSION_WHITELIST} ${PYBIN}";
         continue;
-    elif [[ "${PYBIN}" == *"33"* ]]; then
-        echo "Python 3.3 (via ${PYBIN}) unsupported";
-        echo "=====================================";
-        continue;
-    elif [[ "${PYBIN}" == *"34"* ]]; then
-        echo "Python 3.4 (via ${PYBIN}) unsupported";
-        echo "=====================================";
+    elif [[ "${PYBIN}" == *"36"* ]]; then
+        VERSION_WHITELIST="${VERSION_WHITELIST} ${PYBIN}";
         continue;
     else
-        echo "Building with version: ${PYBIN}";
+        echo "Ignoring unsupported version: ${PYBIN}";
+        echo "=====================================";
     fi
+done
+
+# Compile wheels
+for PYBIN in ${VERSION_WHITELIST}; do
     "${PYBIN}/pip" install -r /io/scripts/manylinux/dev-requirements.txt
     "${PYBIN}/pip" wheel /io/ -w wheelhouse/
 done
@@ -35,4 +36,13 @@ done
 # Bundle external shared libraries into the wheels
 for whl in wheelhouse/${PKG_NAME}*.whl; do
     auditwheel repair "${whl}" -w /io/wheelhouse/
+    rm -f "${whl}"
+done
+
+# Install packages and test
+for PYBIN in ${VERSION_WHITELIST}; do
+    "${PYBIN}/pip" install bezier --no-index \
+        --find-links /io/wheelhouse --find-links wheelhouse
+    (cd "$HOME"; "${PYBIN}/py.test" /io/tests/)
+    (cd "$HOME"; PYTHONPATH=/io/functional_tests/ "${PYBIN}/py.test" /io/functional_tests/)
 done
