### 2017-07-10 b0cd660
Moving curve-curve fnl. test cases into language agnostic JSON files.
Numerical values are represented in one of three ways:

- integers (the only type in JSON is JavaScript's "number" so
  these only get 53 bits of precision rather than 63)
- fractions as strings "a/b"
- IEEE-754 double precision / binary64 numbers as strings (this
  isn't really necessary because JavaScript's "number" uses
  IEEE-754, but this is just extra caution to avoid rounding from
  decimal values to binary values)

Caveat: The files aren't quite language agnostic because the
        IEEE-754 representations come dire
[…]
