### 2017-07-11 0eae9e1
Actually using JSON data in ``test_implicitized_intersect``.
Along the way discovered some typos in ``intersections.json``. Also
had to update the "custom error" for (11, 26) in
``test_implicitized_intersect.py`` because the "best" answer was
even more ULPs away from the computed answer.
