### 2017-07-11 d81b60c
Adding "intersection type" to ``intersections.json``.
Using this in ``test_implicitized_intersect.py`` to switch on
which ``check_*`` function to use.
