### 2017-07-11 47d05ea
Using @pytest.mark.parametrize to run each intersection in ``test_implicitized_intersect.py``.
