+++ functional_tests/test_implicitized_intersect.py
@@ -14,7 +14,6 @@ from __future__ import absolute_import
 
 import numpy as np
 import pytest
-import six
 
 from bezier import _implicitization
 
