### 2017-07-11 c87c9ef
Adding IDs for `parametrize`-d functional tests.
This way, a failure can actually give a human-readable
description of what was being tested.

Big time H/T to:
https://hackebrot.github.io/pytest-tricks/param_id_func/
