### 2017-07-11 7e2d800
Using JSON fnl. test data in `test_curve_curve`.
In the process

- Removing `candidate_curves` module altogether
- (Temporarily) removing `make_plots` helper
- (Temporarily) removing the cases in `test_curve12_self_crossing`
- Added a "do-nothing" context manager in `runtime_utils` that
  can be used side-by-side with the "wiggle" and "raises" ctx. mgrs.
- Moving `id_func` from `test_implicitized_intersect` to
  `runtime_utils`.

Also manually verified all the deleted tests are present as pairs
in `intersections.json`.
