+++ functional_tests/test_curve_curve.py
@@ -14,299 +14,146 @@ from __future__ import absolute_import
 
 import operator
 
-try:
-    import matplotlib.pyplot as plt
-except ImportError:
-    plt = None
 import pytest
 import six
 
+import bezier
 from bezier import _intersection_helpers
-from bezier import _plot_helpers
 
-import candidate_curves
 import runtime_utils
 
 
 CONFIG = runtime_utils.Config()
 S_PROP = operator.attrgetter('s')
+CURVES, INTERSECTIONS = runtime_utils.get_intersections_info()
+WIGGLES = {
+    (21, 22): 12,
+    (11, 26): 25,
+    (8, 27): 42,
+    (32, 33): 4,
+    (50, 54): 91,
+    (51, 54): 1013,
+    (52, 54): 91,
+    (53, 54): 1013,
+}
+FAILURE_NOT_IMPLEMENTED = (
+    (14, 15),  # Line segments parallel.
+    (1, 24),  # The number of candidate intersections is too high. (24)
+    (28, 29),  # The number of candidate intersections is too high. (22)
+)
+INCORRECT_COUNT = (
+    (38, 39),
+)
 
 
-def make_plots(curve1, curve2, points=None,
-               ignore_save=False, failed=True):
-    if not CONFIG.running:
-        return
+class IncorrectCount(ValueError):
+    """Custom exception for a "very bad" answer.
 
-    if points is None:
-        key = (curve1, curve2)
-        info = candidate_curves.INTERSECTION_INFO[key]
-        points = info[:, 2:]
+    This should be raised when the **computed** number of intersections
+    disagrees with the actual number of intersections.
+    """
 
-    if isinstance(curve1, six.integer_types):
-        curve1 = candidate_curves.CURVES[curve1]
-        curve2 = candidate_curves.CURVES[curve2]
 
-    ax = curve1.plot(64)
-    curve2.plot(64, ax=ax)
-    ax.plot(points[:, 0], points[:, 1],
-            marker='o', linestyle='None', color='black')
-    ax.axis('scaled')
-    _plot_helpers.add_plot_boundary(ax)
+def _get_curve(curve_id):
+    curve_info = CURVES[curve_id - 1]
+    assert curve_info['id'] == curve_id
+    return bezier.Curve.from_nodes(
+        curve_info['control_points'], _copy=False)
 
-    if CONFIG.save_plot:
-        if not ignore_save:
-            CONFIG.save_fig()
+
+def _get_params(intersection_info):
+    s_vals = intersection_info['curve1_params']
+    num_s, = s_vals.shape
+
+    t_vals = intersection_info['curve2_params']
+    assert t_vals.shape == (num_s,)
+
+    intersection_pts = intersection_info['intersections']
+    if num_s == 0:
+        assert intersection_pts.size == 0
     else:
-        if failed:
-            plt.title(CONFIG.current_test + ': failed')
-        else:
-            plt.title(CONFIG.current_test)
-        plt.show()
+        assert intersection_pts.shape == (num_s, 2)
 
-    plt.close(ax.figure)
+    return s_vals, t_vals, intersection_pts
 
 
-def curve_curve_check(curve_id1, curve_id2, ignore_save=False):
-    # pylint: disable=too-many-locals
-    key = (curve_id1, curve_id2)
-    # NOTE: This assumes ``info`` is sorted by s-value.
-    info = candidate_curves.INTERSECTION_INFO[key]
-    s_vals = info[:, 0]
-    t_vals = info[:, 1]
-    points = info[:, 2:]
-
-    assert len(s_vals) == len(t_vals)
-    assert len(s_vals) == len(points)
-
[…]
