+++ functional_tests/candidate_curves.py
@@ -1,555 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import numpy as np
-
-import bezier
-
-import runtime_utils
-
-
-CURVES = (
-    None,  # Start counting at one.
-    # g1 = sympy.Matrix([[s, 2 * s * (1 - s)]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 0.0],
-        [0.5, 1.0],
-        [1.0, 0.0],
-    ]), _copy=False),
-    # g2 = sympy.Matrix([[(9 - 8 * s) / 8, (2 * s - 1)**2 / 2]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [1.125, 0.5],
-        [0.625, -0.5],
-        [0.125, 0.5],
-    ]), _copy=False),
-    # g3 = 3 * g1
-    # g3 = sympy.Matrix([[3 * s, 6 * s * (1 - s)]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 0.0],
-        [1.5, 3.0],
-        [3.0, 0.0],
-    ]), _copy=False),
-    # g4 = sympy.Matrix([[
-    #     -3 * (4 * s**2 + s - 4) / 4,
-    #     (92 * s**2 - 77 * s + 24) / 16,
-    # ]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [3.0, 1.5],
-        [2.625, -0.90625],
-        [-0.75, 2.4375],
-    ]), _copy=False),
-    # g5 = sympy.Matrix([[s, (8 * s**2 - 8 * s + 3) / 4]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 0.75],
-        [0.5, -0.25],
-        [1.0, 0.75],
-    ]), _copy=False),
-    # g6 = sympy.Matrix([[s, s**2 + (1 - s)**2]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 1.0],
-        [0.5, 0.0],
-        [1.0, 1.0],
-    ]), _copy=False),
-    # g7 = sympy.Matrix([[s, (4 * s**2 - 4 * s + 17) / 64]])
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 0.265625],
-        [0.5, 0.234375],
-        [1.0, 0.265625],
-    ]), _copy=False),
-    # g8 = sympy.Matrix([[8 * s, 3]]) / 8
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.0, 0.375],
-        [1.0, 0.375],
-    ]), _copy=False),
-    # g9 = sympy.Matrix([[2, 3 * s]]) / 4
-    bezier.Curve.from_nodes(np.asfortranarray([
-        [0.5, 0.0],
[…]
