+++ functional_tests/test_implicitized_intersect.py
@@ -92,16 +92,10 @@ def check_no_intersect(nodes1, nodes2):
     assert param_vals.size == 0
 
 
-def id_func(intersection_info):
-    return 'curves {:d} and {:d} (ID: {:d})'.format(
-        intersection_info['curve1'], intersection_info['curve2'],
-        intersection_info['id'])
-
-
 @pytest.mark.parametrize(
     'intersection_info',
     INTERSECTIONS,
-    ids=id_func,
+    ids=runtime_utils.id_func,
 )
 def test_intersect(intersection_info):
     # Get info for "curve 1".
