### 2017-07-11 78efd1b
Making fnl. test assertions on int. type for "failures".
