### 2017-07-11 4c52dad
Removing `test_curve12_self_crossing.png` image.
It is/was identical to `test_curves42_and_43.png`:

  $ diff -s \
  > docs/images/test_curve12_self_crossing.png \
  > docs/images/test_curves42_and_43.png
  Files docs/images/test_curve12_self_crossing.png and
  docs/images/test_curves42_and_43.png are identical

This is because of the way images were saved in `test_curve_curve`
before 7e2d800423414d9a5eaae62905c58b657a6c07df:

  def test_curve12_self_crossing():
      curve_curve_check(42, 43)
      curve_curve_check(44, 45, ignore_save=True)
      curve_curve_check(46, 47, ignore_save=True)
