### 2017-07-11 3bdfb44
Adding back `make_plots` functionality from `test_curve_curve`.
Put it back as a stand-alone `make_curve_intersection_images`
script that focuses on the test cases and does not require doing
actual intersection for plotting.

Also renaming `intersections.json` to `curve_intersections.json`
since there will also be `surface_intersections.json`
eventually.
