### 2017-07-12 9f401a7
Converting `curves.json` from a list to a dict.
This way the curve IDs are not sensitive to re-ordering.
