### 2017-07-12 0e853e8
Updating `make_curve_intersection_images` to produce "better" image names.
Also updated all files that referenced `test_curvesX_and_Y.png` to
instead reference `curvesX_and_Y.png`.
