+++ docs/curve-curve-intersection.rst
@@ -38,7 +38,7 @@ Curve-Line Intersection
    array([[ 0.25 , 0.375],
           [ 0.75 , 0.375]])
 
-.. image:: images/test_curves1_and_8.png
+.. image:: images/curves1_and_8.png
    :align: center
 
 .. doctest:: intersect-1-9
@@ -58,7 +58,7 @@ Curve-Line Intersection
    >>> curve1.intersect(curve2)
    array([[ 0.5, 0.5]])
 
-.. image:: images/test_curves1_and_9.png
+.. image:: images/curves1_and_9.png
    :align: center
 
 .. doctest:: intersect-10-11
@@ -78,7 +78,7 @@ Curve-Line Intersection
    >>> curve1.intersect(curve2)
    array([[ 3., 4.]])
 
-.. image:: images/test_curves10_and_11.png
+.. image:: images/curves10_and_11.png
    :align: center
 
 .. doctest:: intersect-8-9
@@ -97,7 +97,7 @@ Curve-Line Intersection
    >>> curve1.intersect(curve2)
    array([[ 0.5 , 0.375]])
 
-.. image:: images/test_curves8_and_9.png
+.. image:: images/curves8_and_9.png
    :align: center
 
 .. doctest:: intersect-29-30
@@ -117,7 +117,7 @@ Curve-Line Intersection
    >>> curve1.intersect(curve2)
    array([[ 0., 1.]])
 
-.. image:: images/test_curves29_and_30.png
+.. image:: images/curves29_and_30.png
    :align: center
 
 Curved Intersections
@@ -146,7 +146,7 @@ with zero error:
    array([[ 0.25 , 0.375],
           [ 0.75 , 0.375]])
 
-.. image:: images/test_curves1_and_5.png
+.. image:: images/curves1_and_5.png
    :align: center
 
 .. doctest:: intersect-3-4
@@ -168,7 +168,7 @@ with zero error:
    array([[ 0.75  , 1.125  ],
           [ 2.625 , 0.65625]])
 
-.. image:: images/test_curves3_and_4.png
+.. image:: images/curves3_and_4.png
    :align: center
 
 .. doctest:: intersect-14-16
@@ -190,7 +190,7 @@ with zero error:
    array([[ 0.375 , 0.46875],
           [ 0.625 , 0.46875]])
 
-.. image:: images/test_curves14_and_16.png
+.. image:: images/curves14_and_16.png
    :align: center
 
 Even for curves which don't intersect at exact floating point
@@ -221,7 +221,7 @@ numbers, we can compute the intersection to machine precision:
    >>> binary_exponent(max_err)
    -54
 
-.. image:: images/test_curves1_and_2.png
+.. image:: images/curves1_and_2.png
    :align: center
 
 .. doctest:: intersect-1-7
@@ -249,7 +249,7 @@ numbers, we can compute the intersection to machine precision:
    >>> binary_exponent(max_err)
    -54
 
-.. image:: images/test_curves1_and_7.png
+.. image:: images/curves1_and_7.png
    :align: center
 
 .. doctest:: intersect-1-13
@@ -281,7 +281,7 @@ numbers, we can compute the intersection to machine precision:
    >>> binary_exponent(max_err)
    -54
 
-.. image:: images/test_curves1_and_13.png
+.. image:: images/curves1_and_13.png
    :align: center
 
 .. doctest:: intersect-21-22
@@ -311,7 +311,7 @@ numbers, we can compute the intersection to machine precision:
    >>> binary_exponent(max_err)
    -51
 
-.. image:: images/test_curves21_and_22.png
+.. image:: images/curves21_and_22.png
    :align: center
 
 For higher degree intersections, the error starts to get a little
@@ -345,7 +345,7 @@ larger.
[…]
