### 2017-07-12 5a950b6
Updating note in `make_images` to drop MATPLOTLIBRC env. var.
This is because `nox` now sets this for us.
