### 2017-07-12 6cf19b8
Small tweaks to `WIGGLES` in `test_curve_curve`.
Will likely make an effort to bring `test_curve_curve` and
`test_implicitized_intersect` to use the same method for
equality checking. It makes more sense to use the one from
`test_implicitized_intersect` since it pinpoints the spots
where accuracy fails rather than allowing all values to have
wiggle.
