### 2017-07-12 0beabc6
Adding notes to `surfaces.json`.
In the process

- refining a note on "curve 54", which is contained in "surface 33Q" as
  an edge and then updating the note on "surface 33Q" in
  `test_surface_surface` to match
- fixing a typo in a "note" comment in `test_surface_surface`
