### 2017-07-12 8652d46
Changing "degree elevated" -> "degree-elevated" for consistency.
