### 2017-07-12 8cda98b
Adding `seaborn.set()` (now required as of `0.8`).
