+++ functional_tests/runtime_utils.py
@@ -35,6 +35,8 @@ import six
 from bezier import _helpers
 
 
+if seaborn is not None:
+    seaborn.set()  # Required in `seaborn >= 0.8`
 _FNL_TESTS_DIR = os.path.dirname(__file__)
 _DOCS_DIR = os.path.abspath(
     os.path.join(_FNL_TESTS_DIR, '..', 'docs'))
