### 2017-07-12 fb86303
Making `seaborn.set()` docs stub return `None`.
