### 2017-07-13 0fa0dc7
Fixing "incorrect" using of literal ellipsis for `typing`.
I was under the impression it applied to all iterable containers,
but it is only for `Tuple`. From the docs [1]

  To specify a variable-length tuple of homogeneous type, use
  literal ellipsis, e.g. `Tuple[int, ...]`. A plain `Tuple` is
  equivalent to `Tuple[Any, ...]`, and in turn to `tuple`.

Also using '{:b}' format string instead of using `bin()`.

[1]: https://docs.python.org/3/library/typing.html
