### 2017-07-13 ccfd43f
Make curve intersection runtime utils have "curve"-specific names.
- Renaming `id_func()` to `curve_id_func()` (and adding an
  equivalent `surface_id_func()` even though it isn't used yet)
- Renaming `get_intersections_info()` to `curve_intersections_info()`

Also adding
