### 2017-07-13 aee85f3
Spec-ing out intersection format in `surface_intersections.json`.
- This is "incomplete" because it only considers an intersection
  which is a single `CurvedPolygon` (some surfaces intersect
  in multiple disconnected regions)
- Using `abspath` when setting `FNL_TESTS_DIR` in `runtime_utils`
  because it gets used by functions (it was previously just
  used in setup)
- Updating `convert_floats()` to ignore missing keys
- Update docstring in `curve_intersections_info()` to reflect the
  change in `curves.json` from a list to a dict
- Adding `surface_intersections_info()` to load the info from

[…]
