### 2017-07-14 0eccf6e
Adding another case to `surface_intersections.json`.
In this case, "6Q" and "7Q" intersect in two distinct
regions, so I had to "invent" a way to indicate this in
the data specification. Rather than proliferting more keys
or adding more nested, I chose to use a `null` separator
in the affected data structures.
