### 2017-07-14 137819b
Adding custom `CurveInfo` class for parsing `curves.json`.
This serves not just as a parser but also as a way to document
the content of the ``curves.json`` file.
