### 2017-07-15 7ca4754
Adding some more curve-curve intersections.
One is a "typical" tangency I had used before and the
other is a special tangency: one in which the curvature of
both curves is also equal at the intersection, but the
curves are different.
