### 2017-07-15 480dfdc
Adding CurveIntersectionInfo class and using it **a lot**.
