+++ functional_tests/runtime_utils.py
@@ -15,6 +15,7 @@
 
 import argparse
 import contextlib
+import enum
 import inspect
 import io
 import json
@@ -122,21 +123,6 @@ def no_op_manager():
     yield
 
 
-def curve_id_func(intersection_info):
-    """Turn info from ``curve_intersections.json`` into a test ID.
-
-    Args:
-        intersection_info (dict): An intersection value loaded from
-            ``curve_intersections.json``.
-
-    Returns:
-        str: An identifier formatted from the info.
-    """
-    return 'curves {!r} and {!r} (ID: {:d})'.format(
-        intersection_info['curve1'], intersection_info['curve2'],
-        intersection_info['id'])
-
-
 def surface_id_func(intersection_info):
     """Turn info from ``surface_intersections.json`` into a test ID.
 
@@ -190,9 +176,10 @@ def curve_intersections_info():
 
     filename = os.path.join(FNL_TESTS_DIR, 'curve_intersections.json')
     with io.open(filename, 'r', encoding='utf-8') as file_obj:
-        intersections = json.load(file_obj)
-    keys = ['intersections', 'curve1_params', 'curve2_params']
-    convert_floats(intersections, keys=keys)
+        intersections_json = json.load(file_obj)
+
+    intersections = [CurveIntersectionInfo.from_json(info, curves)
+                     for info in intersections_json]
 
     return curves, intersections
 
@@ -395,9 +382,227 @@ class CurveInfo(object):  # pylint: disable=too-few-public-methods
 
         Returns:
             .CurveInfo: The curve info parsed from the JSON.
+
+        Raises:
+            ValueError: If any of ``info`` is left unparsed.
         """
         control_points = info.pop('control_points')
         control_points = np.asfortranarray(_convert_float(control_points))
         implicitized = info.pop('implicitized', None)
+
+        # Optional fields.
         note = info.pop('note', None)
+
+        if info:
+            raise ValueError('Unexpected keys remaining in JSON info', info)
+
         return cls(id_, control_points, implicitized=implicitized, note=note)
+
+
+class CurveIntersectionType(enum.Enum):
+    """Enum describing curve intersection."""
+
+    coincident = 'coincident'
+    """Curves lie on the same underlying algebraic curve."""
+
+    no_intersection = 'no-intersection'
+    """Curves do not intersect."""
+
+    tangent = 'tangent'
+    """Curves are tangent at the point of intersection."""
+
+    standard = 'standard'
+    """Intersection is not **any** of the other types."""
+
+
+class CurveIntersectionInfo(object):
+    r"""Information about an intersection from ``curve_intersections.json``.
+
+    A curve-curve intersection JSON is expected to have 8 keys:
+
+    * ``curve1``: ID of the first curve in the intersection.
+    * ``curve2``: ID of the second curve in the intersection.
+    * ``id``
+    * ``type``
+    * ``note`` (optional): Description of the intersection / notes about how
+      it behaves.
+    * ``intersections``: List of pairs of "numerical" ``x-y`` values.
+    * ``curve1_params``: "Numerical" curve parameters at intersections.
+    * ``curve1_polys`` (optional): The (integer) coefficients of the minimal
+      polynomials that determine the values in ``curve1_params``.
+
+      For example, if the roots are
+
+      .. math::
+
+         0 \\
+         \frac{7 - \sqrt{7}}{14} \approx \mathtt{0x1.3e7b70cac040dp-2} \\
+         \frac{7 + \sqrt{7}}{14} \approx \mathtt{0x1.60c2479a9fdfap-1} \\
+         1
+
+      Then the corresponding (minimal) polynomial coefficients are:
+
+      .. code-block:: python
+
+         [
[…]

[…]
