### 2017-07-15 cc6c047
Fixing usage of CurveIntersectionInfo in `make_curve_intersection_images`.
Also removing unused `CURVES` global in the 3 modules that call
`curve_intersections_info()`.
