### 2017-07-15 5e328d6
Lint fix for `CurveIntersectionInfo`.
Disabling Pylint warnings for too many instance arguments
(has 11, max is 8) and too many constructor arguments
(has 11, max is 7).
