### 2017-07-15 a012113
Adding `CurveIntersectionInfo._verify_data()`.
