### 2017-07-15 13be05d
Drop usage of id_pair in curve-curve fnl. tests.
(This is not needed since each intersection now has a
unique ID.)
