### 2017-07-15 60a7ed8
Renaming curveN -> curveN_info and adding curveN @property-s.
Also adding `eacute` to `runtime_utils` so we can say
"Bezier".
