### 2017-07-15 ceb7048
Fixing old references to curveN when curveN_info was intended.
Also adding some pylint `disable`-s for `@property`-s which
trigger the `missing-return-type-doc` for some strange reason.
