### 2017-07-15 5bbd6af
Adding ``SurfaceInfo`` for fnl. tests.
