### 2017-07-15 fb8bff5
Adding ``SurfaceIntersectionInfo``.
Also

- removing ``convert_floats()``
- renaming ``curveN`` -> ``curveN_info``
