+++ functional_tests/runtime_utils.py
@@ -127,27 +127,6 @@ def no_op_manager():
     yield
 
 
-def convert_floats(info, keys):
-    """Modify ``info`` in-place to convert strings to floating point numbers.
-
-    Args:
-        info (List[dict]): A list of dictionaries to be modified.
-        keys (List[str]): The keys within each dictionary that contain
-            floating point values to be converted from a "custom" form
-            to native Python ``float`` values.
-    """
-    for element in info:
-        for key in keys:
-            if key not in element:
-                continue
-
-            converted = _convert_float(element[key])
-            if isinstance(converted, list):
-                converted = np.asfortranarray(converted)
-
-            element[key] = converted
-
-
 def curve_intersections_info():
     """Load curve and intersections info from JSON file.
 
@@ -166,7 +145,6 @@ def curve_intersections_info():
     filename = os.path.join(FNL_TESTS_DIR, 'curve_intersections.json')
     with io.open(filename, 'r', encoding='utf-8') as file_obj:
         intersections_json = json.load(file_obj)
-
     intersections = [CurveIntersectionInfo.from_json(info, curves)
                      for info in intersections_json]
 
@@ -190,9 +168,9 @@ def surface_intersections_info():
 
     filename = os.path.join(FNL_TESTS_DIR, 'surface_intersections.json')
     with io.open(filename, 'r', encoding='utf-8') as file_obj:
-        intersections = json.load(file_obj)
-    keys = ['intersections', 'start_params', 'end_params']
-    convert_floats(intersections, keys=keys)
+        intersections_json = json.load(file_obj)
+    intersections = [SurfaceIntersectionInfo.from_json(info, surfaces)
+                     for info in intersections_json]
 
     return surfaces, intersections
 
@@ -615,8 +593,8 @@ class CurveIntersectionInfo(object):
             ValueError: If any of ``info`` is left unparsed.
         """
         id_ = info.pop('id')
-        curve1 = curves[info.pop('curve1')]
-        curve2 = curves[info.pop('curve2')]
+        curve1_info = curves[info.pop('curve1')]
+        curve2_info = curves[info.pop('curve2')]
         type_ = CurveIntersectionType(info.pop('type'))
 
         intersections = np.asfortranarray(
@@ -638,7 +616,7 @@ class CurveIntersectionInfo(object):
             raise ValueError('Unexpected keys remaining in JSON info', info)
 
         return cls(
-            id_, curve1, curve2, type_, intersections,
+            id_, curve1_info, curve2_info, type_, intersections,
             curve1_params, curve2_params,
             curve1_polys=curve1_polys, curve2_polys=curve2_polys, note=note)
 # pylint: enable=too-many-instance-attributes
@@ -698,3 +676,162 @@ class SurfaceInfo(object):  # pylint: disable=too-few-public-methods
             raise ValueError('Unexpected keys remaining in JSON info', info)
 
         return cls(id_, control_points, note=note)
+
+
+class SurfaceIntersectionInfo(object):
+    r"""Information about an intersection from ``surface_intersections.json``.
+
+    A surface-surface intersection JSON is expected to have 10 keys:
+
+    * ``surface1``: ID of the first surface in the intersection.
+    * ``surface2``: ID of the second surface in the intersection.
+    * ``id``
+    * ``note`` (optional): Description of the intersection(s).
+    * ``intersections``(optional, for now): List of pairs of "numerical"
+      ``x-y`` values.
+    * ``start_params``(optional, for now): "Numerical" parameters along edge
+      curves at intersections.
+    * ``start_param_polys`` (optional): The (integer) coefficients of the
+      minimal polynomials that determine the values in ``start_params``.
[…]
