### 2017-07-16 1896e73
Lint fix for `SurfaceIntersectionInfo`.
Disabling Pylint warnings for too many instance arguments
(has 10, max is 8) and too many constructor arguments
(has 11, max is 7).

Also adding some pylint `disable`-s for `@property`-s which
trigger the `missing-return-type-doc` for some strange reason.
