### 2017-07-17 83ada17
Moving `SurfaceIntersectionInfo.from_json` to end of source.
