### 2017-07-17 eaa4a5a
Updating error message for `locate()` methods.
Also adding a note that `locate()` / `evaluate*()` are
(essentially) inverses.

See #35 for context.
