### 2017-07-19 4b01808
Small typo fix: assumes -> assume.
