### 2017-07-20 457429d
Fixing typo in `_resolve_and_add` docstring.
