### 2017-07-20 21ed72e
Switching Apache 2.0 link from http to https.
Also adding LICENSE header to `src/bezier/speedup.f90` and
`src/bezier/_speedup.pyf`.
