### 2017-07-20 cb83162
Replacing all remaining http links with https equivalents.
Also removing unnecessary extra after Amazon link for CAGD book.
