+++ paper.bib
@@ -1,7 +1,7 @@
 %% H/T to: http://www.doi2bib.org
 @article{Farouki1987,
   doi = {10.1016/0167-8396(87)90012-4},
-  url = {http://dx.doi.org/10.1016/0167-8396(87)90012-4},
+  url = {https://dx.doi.org/10.1016/0167-8396(87)90012-4},
   year  = {1987},
   month = {Nov},
   publisher = {Elsevier {BV}},
@@ -14,7 +14,7 @@
 }
 @article{Sederberg1989,
   doi = {10.1016/0010-4485(89)90058-4},
-  url = {http://dx.doi.org/10.1016/0010-4485(89)90058-4},
+  url = {https://dx.doi.org/10.1016/0010-4485(89)90058-4},
   year  = {1989},
   month = {Oct},
   publisher = {Elsevier {BV}},
@@ -27,7 +27,7 @@
 }
 @article{Sederberg1989-alg,
   doi = {10.1016/0010-4485(89)90015-8},
-  url = {http://dx.doi.org/10.1016/0010-4485(89)90015-8},
+  url = {https://dx.doi.org/10.1016/0010-4485(89)90015-8},
   year  = {1989},
   month = {Nov},
   publisher = {Elsevier {BV}},
@@ -40,7 +40,7 @@
 }
 @article{Sederberg1990,
   doi = {10.1016/0010-4485(90)90039-f},
-  url = {http://dx.doi.org/10.1016/0010-4485(90)90039-F},
+  url = {https://dx.doi.org/10.1016/0010-4485(90)90039-F},
   year  = {1990},
   month = {Nov},
   publisher = {Elsevier {BV}},
@@ -53,7 +53,7 @@
 }
 @article{Farouki1991,
   doi = {10.1016/0167-8396(91)90047-f},
-  url = {http://dx.doi.org/10.1016/0167-8396(91)90047-F},
+  url = {https://dx.doi.org/10.1016/0167-8396(91)90047-F},
   year  = {1991},
   month = {Feb},
   publisher = {Elsevier {BV}},
@@ -66,7 +66,7 @@
 }
 @article{Dunavant1985,
   doi = {10.1002/nme.1620210612},
-  url = {http://dx.doi.org/10.1002/nme.1620210612},
+  url = {https://dx.doi.org/10.1002/nme.1620210612},
   year  = {1985},
   month = {Jun},
   publisher = {Wiley-Blackwell},
@@ -79,7 +79,7 @@
 }
 @article{Farouki1988,
   doi = {10.1016/0167-8396(88)90016-7},
-  url = {http://dx.doi.org/10.1016/0167-8396(88)90016-7},
+  url = {https://dx.doi.org/10.1016/0167-8396(88)90016-7},
   year  = {1988},
   month = {Jun},
   publisher = {Elsevier {BV}},
@@ -118,7 +118,7 @@
 }
 @article{Kim1998,
   doi = {10.1016/s0010-4485(98)00052-9},
-  url = {http://dx.doi.org/10.1016/S0010-4485(98)00052-9},
+  url = {https://dx.doi.org/10.1016/S0010-4485(98)00052-9},
   year  = {1998},
   month = {Nov},
   publisher = {Elsevier {BV}},
@@ -131,7 +131,7 @@
 }
 @article{Sederberg1986,
   doi = {10.1016/s0010-4485(86)80013-6},
-  url = {http://dx.doi.org/10.1016/S0010-4485(86)80013-6},
+  url = {https://dx.doi.org/10.1016/S0010-4485(86)80013-6},
   year  = {1986},
   month = {Jan},
   publisher = {Elsevier {BV}},
@@ -144,7 +144,7 @@
 }
 @inproceedings{Roth,
   doi = {10.1109/pccga.2000.883971},
[…]
