### 2017-07-20 83b3229
Change wiggle_interval to return "success" bool instead of raising.
This allows the implicitization approach to use it without having
to use exceptions for flow-control.

Fixes #22.
