+++ src/bezier/_intersection_helpers.py
@@ -1029,6 +1029,8 @@ def _from_linearized_low_level_py(
 
     Raises:
         NotImplementedError: If the segment intersection fails.
+        ValueError: If one of the refined parameter values falls outside
+            the unit interval (with wiggle room).
     """
     # pylint: disable=too-many-locals
     s, t, success = segment_intersection(
@@ -1059,8 +1061,13 @@ def _from_linearized_low_level_py(
     # Perform one step of Newton iteration to refine the computed
     # values of s and t.
     refined_s, refined_t = newton_refine(orig_s, nodes1, orig_t, nodes2)
-    refined_s = _helpers.wiggle_interval(refined_s)
-    refined_t = _helpers.wiggle_interval(refined_t)
+    refined_s, success_s = _helpers.wiggle_interval(refined_s)
+    refined_t, success_t = _helpers.wiggle_interval(refined_t)
+    if not (success_s and success_t):
+        raise ValueError(
+            'At least one value outside of unit interval',
+            refined_s, refined_t)
+
     return refined_s, refined_t, True
     # pylint: enable=too-many-locals
 # pylint: enable=too-many-arguments,too-many-return-statements
