### 2017-07-20 ee858ff
Factoring out parameter "wiggle" from `_from_linearized_low_level_py()`.
This was to make it easier to test.

Also fixing a lint issue in `test__helpers.py` (used two spaces).
