### 2017-07-20 b06c78e
Using Fortran-contiguous arrays in `_check_non_simple()`.
