### 2017-07-20 9d47764
Updating "pip install" commands.
