### 2017-07-21 7da2f4c
Adding (way too much) code for parsing `surface_intersections.json`.
