+++ functional_tests/runtime_utils.py
@@ -169,7 +169,7 @@ def surface_intersections_info():
     filename = os.path.join(FNL_TESTS_DIR, 'surface_intersections.json')
     with io.open(filename, 'r', encoding='utf-8') as file_obj:
         intersections_json = json.load(file_obj)
-    intersections = [SurfaceIntersectionInfo.from_json(info, surfaces)
+    intersections = [SurfaceIntersectionsInfo.from_json(info, surfaces)
                      for info in intersections_json]
 
     return surfaces, intersections
@@ -605,7 +605,7 @@ class CurveIntersectionInfo(object):
             _convert_float(info.pop('curve2_params')))
 
         if intersections.size == 0:
-            intersections = intersections.reshape((0, 2))
+            intersections = intersections.reshape((0, 2), order='F')
 
         # Optional fields.
         curve1_polys = info.pop('curve1_polys', None)
@@ -678,8 +678,116 @@ class SurfaceInfo(object):  # pylint: disable=too-few-public-methods
         return cls(id_, control_points, note=note)
 
 
-# pylint: disable=too-many-instance-attributes
 class SurfaceIntersectionInfo(object):
+    """Information about a single curved polygon intersection.
+
+    When two surfaces are intersected, the intersection region may split
+    into multiple disjoint curved polygons (e.g. "6Q"-"7Q").
+
+    Args:
+        intersections (Optional[numpy.ndarray]): ``Nx2`` array of ``x-y``
+            coordinate pairs of intersection points.
+        edge_pairs (List[List[int]]): List of pairs of
+            ``surface_index, edge_index`` for each intersection. I.e. the
+            intersection occurs on ``surface_index`` (one of 1 or 2) and on
+            the edge ``edge_index`` (one of 0, 1 or 2) **of that surface**.
+            An intersection requires **two** edges, but this one is the edge
+            that the boundary (of the intersection) continues along.
+        start_params (numpy.ndarray): 1D array, the parameters along
+            ``surface1`` where the intersections occur (in the same order as
+            the rows of ``intersections``). These are typically called
+            ``s``-parameters.
+        end_params (numpy.ndarray): 1D array, the parameters along
+            ``surface2`` where the intersections occur (in the same order as
+            the rows of ``intersections``). These are typically called
+            ``t``-parameters.
+        start_param_polys (Optional[List[List[int]]]): The coefficients of the
+            polynomials that determine the values in ``start_params``.
+        end_param_polys (Optional[List[List[int]]]): The coefficients of the
+            polynomials that determine the values in ``end_params``.
+    """
+
+    def __init__(self, parent, intersections, edge_pairs,
+                 start_params, end_params,
+                 start_param_polys=None, end_param_polys=None):
+        self.parent = parent
+
+        self.intersections = intersections
+        self.edge_pairs = edge_pairs
+
+        self.start_params = start_params
+        self.start_param_polys = start_param_polys
+
+        self.end_params = end_params
+        self.end_param_polys = end_param_polys
+
+        self.num_intersections = self._verify()
+
+    def _verify_polynomials(self, num_intersections, start=True):
+        """Verify a list of polynomial coefficients.
+
+        Args:
+            num_intersections (int): The (expected) number of intersections.
+            start (Optional[bool]): Indicates if start or end polynomials
+                should be verified.
+
+        Raises:
+            ValueError: If the sizes are not as expected.
+            ValueError: If the types are not as expected.
+        """
+        if start:
+            polynomials = self.start_param_polys
+            name = 'start'
+        else:
+            polynomials = self.end_param_polys
+            name = 'end'
+
+        err_msg1 = 'Unexpected number of {} parameter polynomials.'.format(
+            name)
+        template2 = (
[…]
