### 2017-07-21 5b7d348
Lint fix for runtime_utils "data-mostly" classes.
They make PyLint angry when they hold too much data.
