+++ functional_tests/test_surface_surface.py
@@ -11,8 +11,8 @@
 # limitations under the License.
 
 from __future__ import absolute_import
-
 import collections
+import operator
 
 try:
     import matplotlib.pyplot as plt
@@ -32,12 +32,18 @@ import runtime_utils
 
 ALGEBRAIC = curve.IntersectionStrategy.algebraic
 GEOMETRIC = curve.IntersectionStrategy.geometric
+_, INTERSECTIONS = runtime_utils.surface_intersections_info()
 STRATEGY = GEOMETRIC
 PARALLEL_FAILURE = ('Line segments parallel.',)
 BAD_TANGENT = (
     'Curves moving in opposite direction but define '
     'overlapping arcs.')
 TANGENT_FAILURE = 'The number of candidate intersections is too high.'
+WIGGLES = {
+    1: 48,
+    13: 25,
+}
+WHITELIST = (1, 13)  # This is temporary.
 CONFIG = runtime_utils.Config()
 
 # F1L = sympy.Matrix([[s, t]])
@@ -176,30 +182,6 @@ SURFACE5Q = bezier.Surface.from_nodes(np.asfortranarray([
     [0.4375, 0.53125],
     [0.875, 0.75],
     [0.625, 1.0],
-]), _copy=False)
-# F6Q = sympy.Matrix([[
-#     (s**2 + s * t + 3 * s + 2 * t) / 4,
-#     -(3 * s**2 + 3 * s * t - 3 * s - 4 * t) / 4,
-# ]])
-SURFACE6Q = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.0, 0.0],
-    [0.375, 0.375],
-    [1.0, 0.0],
-    [0.25, 0.5],
-    [0.75, 0.5],
-    [0.5, 1.0],
-]), _copy=False)
-# F7Q = sympy.Matrix([[
-#     -(s**2 + s * t + 3 * s + 2 * t - 4) / 4,
-#     (8 * s**2 + 8 * s * t - 8 * s - 9 * t + 3) / 8,
-# ]])
-SURFACE7Q = bezier.Surface.from_nodes(np.asfortranarray([
-    [1.0, 0.375],
-    [0.625, -0.125],
-    [0.0, 0.375],
-    [0.75, -0.1875],
-    [0.25, -0.1875],
-    [0.5, -0.75],
 ]), _copy=False)
 # F8Q = sympy.Matrix([[
 #     s * (7 * t + 1),
@@ -637,37 +619,6 @@ def surface_surface_check_multi(surface1, surface2, *all_intersected):
     # pylint: enable=too-many-locals
 
 
-def test_surfaces1Q_and_3Q():
-    _, s_val1 = runtime_utils.real_roots([36, -48, 4, 200, -131])
-    s_val2, = runtime_utils.real_roots([49, 91, 320, -244])
-
-    t_val1, _ = runtime_utils.real_roots([9, -18, 5, -28, 12])
-    t_val2, = runtime_utils.real_roots([49, 63, 88, -128])
-    start_vals = np.asfortranarray([0.0, t_val1, s_val2, 0.0])
-    end_vals = np.asfortranarray([s_val1, t_val2, 1.0, 1.0])
-
-    x_val1 = 0.5 * (1.0 - s_val1) * (s_val1 + 2.0)
-    y_val1 = 0.5 * s_val1 * (3.0 - s_val1)
-    x_val2 = 0.5 * s_val2 * (1.0 - s_val2)
-    y_val2 = 1.0 - s_val2
-    nodes = np.asfortranarray([
-        [1.0, 0.0],
-        [x_val1, y_val1],
-        [x_val2, y_val2],
-        [0.0, 0.0],
-    ])
-    edge_pairs = (
[…]
