### 2017-07-21 8ece202
Removing hardcoded surface definitions in `test_surface_surface`.
