+++ functional_tests/test_surface_surface.py
@@ -32,7 +32,7 @@ import runtime_utils
 
 ALGEBRAIC = curve.IntersectionStrategy.algebraic
 GEOMETRIC = curve.IntersectionStrategy.geometric
-_, INTERSECTIONS = runtime_utils.surface_intersections_info()
+SURFACES, INTERSECTIONS = runtime_utils.surface_intersections_info()
 STRATEGY = GEOMETRIC
 PARALLEL_FAILURE = ('Line segments parallel.',)
 BAD_TANGENT = (
@@ -46,444 +46,48 @@ WIGGLES = {
 WHITELIST = (1, 13)  # This is temporary.
 CONFIG = runtime_utils.Config()
 
-# F1L = sympy.Matrix([[s, t]])
-SURFACE1L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.0, 0.0],
-    [1.0, 0.0],
-    [0.0, 1.0],
-]), _copy=False)
-# F2L = sympy.Matrix([[
-#     (9 * s + 2 * t - 1) / 8,
-#     (9 * s + 7 * t - 1) / 16,
-# ]])
-SURFACE2L = bezier.Surface.from_nodes(np.asfortranarray([
-    [-0.125, -0.0625],
-    [1.0, 0.5],
-    [0.125, 0.375],
-]), _copy=False)
-# F3L = sympy.Matrix([[(4 * s + 1) / 4, (8 * t - 1) / 8]])
-SURFACE3L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.25, -0.125],
-    [1.25, -0.125],
-    [0.25, 0.875],
-]), _copy=False)
-# F4L = sympy.Matrix([[5 * (2 * s + t - 1) / 4, 35 * t / 16]])
-SURFACE4L = bezier.Surface.from_nodes(np.asfortranarray([
-    [-1.25, 0.0],
-    [1.25, 0.0],
-    [0.0, 2.1875],
-]), _copy=False)
-# F5L = sympy.Matrix([[
-#     (21 * s - 21 * t + 4) / 8,
-#     (21 * s + 21 * t - 5) / 32,
-# ]])
-SURFACE5L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.5, -0.15625],
-    [3.125, 0.5],
-    [-2.125, 0.5],
-]), _copy=False)
-# F6L = sympy.Matrix([[(1 - 4 * s) / 4, (1 - 4 * t) / 4]])
-SURFACE6L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.25, 0.25],
-    [-0.75, 0.25],
-    [0.25, -0.75],
-]), _copy=False)
-# F7L = sympy.Matrix([[-(s + t), (s - t + 1) / 2]])
-SURFACE7L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.0, 0.5],
-    [-1.0, 1.0],
-    [-1.0, 0.0],
-]), _copy=False)
-# F8L = sympy.Matrix([[
-#     (5 * s + 9 * t + 1) / 8,
-#     -(5 * s - 4 * t - 6) / 8,
-# ]])
-SURFACE8L = bezier.Surface.from_nodes(np.asfortranarray([
-    [0.125, 0.75],
-    [0.75, 0.125],
-    [1.25, 1.25],
-]), _copy=False)
-# F9L = sympy.Matrix([[(7 * s + 2 * t) / 8, -(s - 5 * t - 1) / 8]])
-SURFACE9L = bezier.Surface.from_nodes(np.asfortranarray([
[…]
