### 2017-07-21 3e0c8dd
Adding data for surf-surf IDs 32 and 33.
In the process also fixing some parse / verify issues with
optional polynomials.
