### 2017-07-21 1711f04
Adding "unfinished" notes for all surf-surf test cases that need data.
Then switching to use this note (instead of a WHITELIST) when
skipping tests in `test_surface_surface.py`.
