### 2017-07-22 78a9aac
Adding data for 20Q-21Q surf-surf fnl. test.
In the process

- Changing `check_tangent()` to take the caught exception rather
  than the `pytest` exception info
- Making a `check_tangent_manager()` context manager that **catches**
  an exception and then calls `check_tangent()`
