+++ functional_tests/test_surface_surface.py
@@ -12,6 +12,7 @@
 
 from __future__ import absolute_import
 import collections
+import contextlib
 import operator
 
 try:
@@ -38,12 +39,16 @@ PARALLEL_FAILURE = ('Line segments parallel.',)
 BAD_TANGENT = (
     'Curves moving in opposite direction but define '
     'overlapping arcs.')
+BAD_TANGENT = (BAD_TANGENT,)
 TANGENT_FAILURE = 'The number of candidate intersections is too high.'
 WIGGLES = {
     1: 46,
     13: 19,
     32: 1013,
     33: 1013,
+}
+FAILED_CASES_TANGENT = {
+    10: {'parallel': True},
 }
 CONFIG = runtime_utils.Config()
 
@@ -74,8 +79,6 @@ SURFACE16Q = SURFACES['16Q'].surface
 SURFACE17Q = SURFACES['17Q'].surface
 SURFACE18Q = SURFACES['18Q'].surface
 SURFACE19Q = SURFACES['19Q'].surface
-SURFACE20Q = SURFACES['20Q'].surface
-SURFACE21Q = SURFACES['21Q'].surface
 SURFACE22Q = SURFACES['22Q'].surface
 SURFACE23Q = SURFACES['23Q'].surface
 SURFACE24Q = SURFACES['24Q'].surface
@@ -155,17 +158,31 @@ def curved_polygon_edges(intersection, edges):
     return edge_list[index:] + edge_list[:index]
 
 
-def check_tangent(exc_info, parallel=False, bad_tangent=False):
+def check_tangent(caught_exc, parallel=False, bad_tangent=False):
+    exc_args = caught_exc.args
     if STRATEGY is GEOMETRIC:
         if parallel:
-            assert exc_info.value.args == PARALLEL_FAILURE
+            assert exc_args == PARALLEL_FAILURE
         elif bad_tangent:
-            assert exc_info.value.args == (BAD_TANGENT,)
+            assert exc_args == BAD_TANGENT
         else:
-            assert str(exc_info.value).startswith(TANGENT_FAILURE)
+            assert len(exc_args) == 1
+            assert exc_args[0].startswith(TANGENT_FAILURE)
     else:
-        assert len(exc_info.value.args) == 2
-        assert exc_info.value.args[0] == _implicitization._NON_SIMPLE_ERR
+        assert len(exc_args) == 2
+        assert exc_args[0] == _implicitization._NON_SIMPLE_ERR
+
+
+@contextlib.contextmanager
+def check_tangent_manager(**kwargs):
+    caught_exc = None
+    try:
+        yield
+    except NotImplementedError as exc:
+        caught_exc = exc
+
+    assert caught_exc is not None
+    check_tangent(caught_exc, **kwargs)
 
 
 def check_coincident(exc_info, parallel=False):
@@ -360,7 +377,7 @@ def test_surfaces3Q_and_4Q():
             surface_surface_check(SURFACE3Q, SURFACE4Q,
                                   start_vals, end_vals, nodes, edge_pairs)
 
-        check_tangent(exc_info)
+        check_tangent(exc_info.value)
 
 
 def test_surfaces1Q_and_5L():
@@ -386,7 +403,7 @@ def test_surfaces1Q_and_5L():
         surface_surface_check(SURFACE1Q, SURFACE5L,
                               start_vals, end_vals, nodes, edge_pairs)
 
-    check_tangent(exc_info)
+    check_tangent(exc_info.value)
     intersection = make_curved_polygon(
         SURFACE1Q, SURFACE5L,
         start_vals, end_vals, edge_pairs)
@@ -444,33 +461,6 @@ def test_surfaces1L_and_2L():
[…]
