### 2017-07-22 d455e11
Refactoring part of `SurfaceIntersectionsInfo.from_json()`.
This is the part that handles the (optional) start/end parameter
polynomials.

Worth noting: this code is already out of hand, I need to change
the data structure in `surface_intersections.json`.
