### 2017-07-22 3609eec
Adding 26Q-27Q surf-surf fnl. test to JSON.
