### 2017-07-22 baf7454
Adding surf-surf fnl. tests to JSON for cases that don't intersect.
- 10Q-19Q
- 4Q-10Q
- 1L-7L
- 8L-29Q
- 1L-10L

This has (temporarily) broken the 4Q-10Q case for algebraic intersection,
but it "works for now" (since only geometric is used on CI).
