### 2017-07-22 6273df2
Adding 3Q-13Q surf-surf fnl. tests to JSON.
This has (temporarily) broken the for algebraic intersection,
but it "works for now" (since only geometric is used on CI).
