+++ functional_tests/test_surface_surface.py
@@ -75,7 +75,6 @@ SURFACE10Q = SURFACES['10Q'].surface
 SURFACE14Q = SURFACES['14Q'].surface
 SURFACE17Q = SURFACES['17Q'].surface
 SURFACE19Q = SURFACES['19Q'].surface
-SURFACE28Q = SURFACES['28Q'].surface
 SURFACE29Q = SURFACES['29Q'].surface
 
 
@@ -486,26 +485,6 @@ def test_surfaces1L_and_6L():
         (0, 2),
     )
     surface_surface_check(SURFACE1L, SURFACE6L,
-                          start_vals, end_vals, nodes, edge_pairs)
-
-
-def test_surfaces1L_and_28Q():
-    _, s_val3 = runtime_utils.real_roots([5, 30, -13])
-    t_val3, _ = runtime_utils.real_roots([5, -40, 22])
-    start_vals = np.asfortranarray([0.1875, 0.0, t_val3])
-    end_vals = np.asfortranarray([1.0, s_val3, 1.0])
-
-    nodes = np.asfortranarray([
-        [0.1875, 0.0],
-        [1.0, 0.0],
-        [1.0 - s_val3, s_val3],
-    ])
-    edge_pairs = (
-        (0, 0),
-        (0, 1),
-        (1, 1),
-    )
-    surface_surface_check(SURFACE1L, SURFACE28Q,
                           start_vals, end_vals, nodes, edge_pairs)
 
 
