### 2017-07-22 a8bf7b3
Adding 3Q-4Q surf-surf fnl. tests to JSON.
This has (temporarily) broken this case for algebraic intersection,
but it "works for now" (since only geometric is used on CI).
