### 2017-07-22 7209c7c
Adding 1L-6L surf-surf fnl. tests to JSON.
