### 2017-07-22 b832329
Adding 1L-9L surf-surf fnl. tests to JSON.
This concludes the migration of test cases into JSON.
