### 2017-07-22 c8e02d2
Removing unused import and fn. in `test_surface_surface`.
