### 2017-07-22 d257fa6
Collapsing check_*_manager() code in `test_surface_surface`.
The two functions `check_tangent` and `check_coincident` no longer
have any other callers, so it's not worth keeping them separate.
