### 2017-07-22 2d1d170
Removing unused `make_curved_polygon`.
