### 2017-07-22 f606ba8
Making strategy a local/argument instead of a global in `test_surface_surface`.
This is so we can make `strategy` a pytest parameter.
