### 2017-07-22 17e64b6
Making strategy a pytest parameter in `test_surface_surface`.
