### 2017-07-22 1793589
Adding in algebraic strategyin `test_surface_surface`.
To do this, made a "first-level" of `strategy` keys in the
`WIGGLES`, `FAILED_CASES_TANGENT` and `FAILED_CASES_COINCIDENT`.
