### 2017-07-22 eaf11d4
Removing `Intersected` fnl. test data container.
