### 2017-07-22 a3654ba
Renaming `make_curve_intersection_images.py`.
