### 2017-07-22 fc9885c
Renaming `surface_surface_check_multi()` -> `surface_surface_check()`.
The distinction between "single" and "multi" is no longer
needed after recent code changes.
