### 2017-07-22 0f1be2e
Adding `make_surface_surface_images` script.
In the process, removing any plotting from
`test_surface_surface`.
