+++ functional_tests/test_surface_surface.py
@@ -86,34 +86,6 @@ FAILED_CASES_COINCIDENT = {
 CONFIG = runtime_utils.Config()
 
 
-def make_plots(surface1, surface2, intersections, failed=True):
-    if not CONFIG.running:
-        return
-
-    ax = surface1.plot(64)
-    surface2.plot(64, ax=ax)
-
-    color = None
-    for intersection in intersections:
-        intersection.plot(64, color=color, ax=ax)
-        # Color is (R,G,B,A) but we just want (R,G,B).
-        color = ax.patches[-1].get_facecolor()[:3]
-
-    plt.axis('scaled')
-    _plot_helpers.add_plot_boundary(ax)
-
-    if CONFIG.save_plot:
-        CONFIG.save_fig()
-    else:
-        if failed:
-            plt.title(CONFIG.current_test + ': failed')
-        else:
-            plt.title(CONFIG.current_test)
-        plt.show()
-
-    plt.close(ax.figure)
-
-
 def curved_polygon_edges(intersection, edges):
     edges1, edges2 = edges
     all_edges = edges1 + edges2
@@ -206,8 +178,6 @@ def surface_surface_check(strategy, surface1, surface2, *all_intersected):
             CONFIG.assert_close(edge.end, end_val)
             CONFIG.assert_close(edge._nodes[0, 0], node[0])
             CONFIG.assert_close(edge._nodes[0, 1], node[1])
-
-    make_plots(surface1, surface2, intersections, failed=False)
     # pylint: enable=too-many-locals
 
 
