### 2017-07-22 505f9af
Dropping unused global `SURFACES`.
