### 2017-07-22 b87a04c
Running `make_surface_surface_images.py` for the first time.
- Adding "new" images
- Deleting old `test_surfacesX_and_Y.png` images
- Handling any renamed image links in docs
