### 2017-07-22 ceacf95
Removing unused imports in `test_surface_surface.py`.
