### 2017-07-22 7b4385c
Renaming `SurfaceIntersectionInfo` -> `CurvedPolygonInfo`.
