### 2017-07-23 cb008d7
Moving "parse" complexity into data structure in `surface_intersections.json`.
(This should've been done already.)
