+++ functional_tests/runtime_utils.py
@@ -351,9 +351,6 @@ class CurveInfo(object):  # pylint: disable=too-few-public-methods
 
         Returns:
             .CurveInfo: The curve info parsed from the JSON.
-
-        Raises:
-            ValueError: If any of ``info`` is left unparsed.
         """
         control_points = info.pop('control_points')
         control_points = np.asfortranarray(_convert_float(control_points))
@@ -361,9 +358,7 @@ class CurveInfo(object):  # pylint: disable=too-few-public-methods
 
         # Optional fields.
         note = info.pop('note', None)
-
-        if info:
-            raise ValueError('Unexpected keys remaining in JSON info', info)
+        _ensure_empty(info)
 
         return cls(id_, control_points, implicitized=implicitized, note=note)
 
@@ -588,9 +583,6 @@ class CurveIntersectionInfo(object):
 
         Returns:
             .CurveIntersectionInfo: The intersection info parsed from the JSON.
-
-        Raises:
-            ValueError: If any of ``info`` is left unparsed.
         """
         id_ = info.pop('id')
         curve1_info = curves[info.pop('curve1')]
@@ -611,9 +603,7 @@ class CurveIntersectionInfo(object):
         curve1_polys = info.pop('curve1_polys', None)
         curve2_polys = info.pop('curve2_polys', None)
         note = info.pop('note', None)
-
-        if info:
-            raise ValueError('Unexpected keys remaining in JSON info', info)
+        _ensure_empty(info)
 
         return cls(
             id_, curve1_info, curve2_info, type_, intersections,
@@ -662,18 +652,13 @@ class SurfaceInfo(object):  # pylint: disable=too-few-public-methods
 
         Returns:
             .SurfaceInfo: The surface info parsed from the JSON.
-
-        Raises:
-            ValueError: If any of ``info`` is left unparsed.
         """
         control_points = info.pop('control_points')
         control_points = np.asfortranarray(_convert_float(control_points))
 
         # Optional fields.
         note = info.pop('note', None)
-
-        if info:
-            raise ValueError('Unexpected keys remaining in JSON info', info)
+        _ensure_empty(info)
 
         return cls(id_, control_points, note=note)
 
@@ -686,10 +671,34 @@ class CurvedPolygonInfo(object):
     because when two surfaces are intersected, the intersection region may
     split into multiple disjoint curved polygons (e.g. "6Q"-"7Q").
 
+    Curved polygon JSON coming from a surface-surface intersection is expected
+    to have 6 keys:
+
+    * ``nodes``: List of pairs of "numerical" ``x-y`` values.
+    * ``start_params``: "Numerical" parameters along edge curves at
+      intersections.
+    * ``start_param_polys`` (optional): The (integer) coefficients of the
+      minimal polynomials that determine the values in ``start_params``.
+      See ``curve1_params`` in :class:`CurveIntersectionInfo` for an
+      example.
+    * ``end_params``: "Numerical" parameters along edge curves at
+      intersections.
+    * ``end_param_polys`` (optional): The (integer) coefficients of the
+      minimal polynomials that determine the values in ``end_params``.
+    * ``edge_pairs``: List of pairs of ``surface_index, edge_index`` for
+      each intersection, i.e. the intersection occurs on ``surface_index``
+      (one of 1 or 2) and on the edge ``edge_index`` (one of 0, 1 or 2) **of
+      that surface**. An intersection requires **two** edges, but this one is
+      the edge that the boundary (of the intersection) continues along.
+
+    The "numerical" values in ``nodes``, ``start_params`` and
+    ``end_params`` can be integers, stringified fractions or stringified
+    IEEE-754 values (``%a`` format).
+
[…]
