### 2017-07-23 047f334
Updating `DEVELOPMENT` doc with instructions about doc images.
