### 2017-07-23 a669d8b
Small tweak to `index.rst` in Sphinx docs.
This is to make it more similar to `README.rst`. (I may try to
generalize `README.rst.template` to cover both so they
don't get too out of sync.)
