### 2017-07-24 8425deb
Adding note in DEVELOPMENT about JSON fnl. test data.
Fixes #34.
