### 2017-07-24 a14237a
Adding JSON-schema for curve-curve intersection fnl. test cases.
