### 2017-07-24 f5a77ea
Slight refactor to `CurveIntersectionInfo._verify_dimensions`.
