### 2017-07-24 8bab4b2
Adding JSON-schema for `surfaces.json`.
