### 2017-07-24 6ae9137
Adding JSON-schema for curved polygons in surf-surf intersections.
