### 2017-07-24 14c0e03
Adding JSON-schema for `surface_intersections.json`.
Also adding note about JSON schema in `DEVELOPMENT.rst`.
