### 2017-07-24 5e48b16
Moving `id_func()` into `runtime_utils`.
