### 2017-07-24 bd87a61
Making strategy a parameter in `test_curve_curve`.
This is in preparation for unifying `test_curve_curve` and
`test_implicitized_intersect`.
