### 2017-07-24 fa81018
Re-writing `test_curve_curve` to be more like `test_implicitized_intersect`.
In particular, setting a "wiggle" threshold on a per-value basis rather than
setting it for the entire case. Had to add some extra support here (that
wasn't needed in algebraic case) for values that have some round-off error
when approximating zero.
