+++ functional_tests/test_curve_curve.py
@@ -15,6 +15,7 @@ from __future__ import absolute_import
 import itertools
 import operator
 
+import numpy as np
 import pytest
 import six
 
@@ -25,20 +26,61 @@ import runtime_utils
 from runtime_utils import CurveIntersectionType
 
 
+SPACING = np.spacing  # pylint: disable=no-member
 CONFIG = runtime_utils.Config()
 GEOMETRIC = bezier.curve.IntersectionStrategy.geometric
 S_PROP = operator.attrgetter('s')
 _, INTERSECTIONS = runtime_utils.curve_intersections_info()
-WIGGLES = {
-    12: 4,  # Established on Ubuntu 16.04 (Less than 8)
-    17: 7,  # Established on Ubuntu 16.04 (Less than 8)
-    18: 11,  # Established on Ubuntu 16.04
-    23: 41,  # Established on Ubuntu 16.04
-    28: 3,  # Established on Ubuntu 16.04 (Less than 8)
-    37: 91,  # Established on Ubuntu 16.04
-    38: 1013,  # Established on Ubuntu 16.04
-    39: 91,  # Established on Ubuntu 16.04
-    40: 1013,  # Established on Ubuntu 16.04
+ULPS_ALLOWED = 3.0
+# NOTE: Set a threshold for values that are "approximately" zero.
+#       This is an **absolute** error rather than a relative error
+#       since relative to zero error is always infinite.
+ZERO_THRESHOLD = 0.5**52
+ZERO_MISS = object()
+ULPS_ALLOWED_OVERRIDE = {
+    12: {
+        (0, 1): 4,  # Established on Ubuntu 16.04
+    },
+    17: {
+        (0, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+        (2, 1): 7,  # Established on Ubuntu 16.04
+        (3, 1): 4,  # Established on Ubuntu 16.04
+        (3, 2): 7,  # Established on Ubuntu 16.04
+    },
+    18: {
+        (0, 6): 5,  # Established on Ubuntu 16.04
+        (1, 0): 8,  # Established on Ubuntu 16.04
+        (1, 2): 10,  # Established on Ubuntu 16.04
+        (1, 3): 11,  # Established on Ubuntu 16.04
+        (1, 6): 4,  # Established on Ubuntu 16.04
+        (2, 2): 4,  # Established on Ubuntu 16.04
+        (3, 3): ZERO_MISS,  # Established on Ubuntu 16.04
+        (3, 5): ZERO_MISS,  # Established on Ubuntu 16.04
+        (3, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+    },
+    23: {
+        (0, 0): 14,  # Established on Ubuntu 16.04
+        (0, 1): 41,  # Established on Ubuntu 16.04
+        (0, 2): 14,  # Established on Ubuntu 16.04
+        (1, 0): 16,  # Established on Ubuntu 16.04
+        (1, 1): 21,  # Established on Ubuntu 16.04
+        (1, 2): 16,  # Established on Ubuntu 16.04
+    },
+    25: {
+        (0, 6): ZERO_MISS,  # Established on Ubuntu 16.04
+    },
+    37: {
+        (0, 1): 91,  # Established on Ubuntu 16.04
+    },
+    38: {
+        (0, 1): 1013,  # Established on Ubuntu 16.04
+    },
+    39: {
+        (0, 1): 91,  # Established on Ubuntu 16.04
+    },
+    40: {
+        (0, 1): 1013,  # Established on Ubuntu 16.04
+    },
 }
 FAILURE_NOT_IMPLEMENTED = (
     11,  # Line segments parallel.
@@ -63,8 +105,7 @@ class IncorrectCount(ValueError):
[…]
