### 2017-07-24 9fb1d37
Scoping the "exceptions-to-the-rule" in `test_curve_curve` to the GEOMETRIC strategy.
