+++ functional_tests/test_curve_curve.py
@@ -38,63 +38,69 @@ ULPS_ALLOWED = 3.0
 ZERO_THRESHOLD = 0.5**52
 ZERO_MISS = object()
 ULPS_ALLOWED_OVERRIDE = {
-    12: {
-        (0, 1): 4,  # Established on Ubuntu 16.04
-    },
-    17: {
-        (0, 7): ZERO_MISS,  # Established on Ubuntu 16.04
-        (2, 1): 7,  # Established on Ubuntu 16.04
-        (3, 1): 4,  # Established on Ubuntu 16.04
-        (3, 2): 7,  # Established on Ubuntu 16.04
-    },
-    18: {
-        (0, 6): 5,  # Established on Ubuntu 16.04
-        (1, 0): 8,  # Established on Ubuntu 16.04
-        (1, 2): 10,  # Established on Ubuntu 16.04
-        (1, 3): 11,  # Established on Ubuntu 16.04
-        (1, 6): 4,  # Established on Ubuntu 16.04
-        (2, 2): 4,  # Established on Ubuntu 16.04
-        (3, 3): ZERO_MISS,  # Established on Ubuntu 16.04
-        (3, 5): ZERO_MISS,  # Established on Ubuntu 16.04
-        (3, 7): ZERO_MISS,  # Established on Ubuntu 16.04
-    },
-    23: {
-        (0, 0): 14,  # Established on Ubuntu 16.04
-        (0, 1): 41,  # Established on Ubuntu 16.04
-        (0, 2): 14,  # Established on Ubuntu 16.04
-        (1, 0): 16,  # Established on Ubuntu 16.04
-        (1, 1): 21,  # Established on Ubuntu 16.04
-        (1, 2): 16,  # Established on Ubuntu 16.04
-    },
-    25: {
-        (0, 6): ZERO_MISS,  # Established on Ubuntu 16.04
-    },
-    37: {
-        (0, 1): 91,  # Established on Ubuntu 16.04
-    },
-    38: {
-        (0, 1): 1013,  # Established on Ubuntu 16.04
-    },
-    39: {
-        (0, 1): 91,  # Established on Ubuntu 16.04
-    },
-    40: {
-        (0, 1): 1013,  # Established on Ubuntu 16.04
+    GEOMETRIC: {
+        12: {
+            (0, 1): 4,  # Established on Ubuntu 16.04
+        },
+        17: {
+            (0, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+            (2, 1): 7,  # Established on Ubuntu 16.04
+            (3, 1): 4,  # Established on Ubuntu 16.04
+            (3, 2): 7,  # Established on Ubuntu 16.04
+        },
+        18: {
+            (0, 6): 5,  # Established on Ubuntu 16.04
+            (1, 0): 8,  # Established on Ubuntu 16.04
+            (1, 2): 10,  # Established on Ubuntu 16.04
+            (1, 3): 11,  # Established on Ubuntu 16.04
+            (1, 6): 4,  # Established on Ubuntu 16.04
+            (2, 2): 4,  # Established on Ubuntu 16.04
+            (3, 3): ZERO_MISS,  # Established on Ubuntu 16.04
+            (3, 5): ZERO_MISS,  # Established on Ubuntu 16.04
+            (3, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+        },
+        23: {
+            (0, 0): 14,  # Established on Ubuntu 16.04
+            (0, 1): 41,  # Established on Ubuntu 16.04
+            (0, 2): 14,  # Established on Ubuntu 16.04
+            (1, 0): 16,  # Established on Ubuntu 16.04
[…]
