### 2017-07-24 df456d4
Re-writing `test_curve_curve` to branch on intersection type.
