### 2017-07-24 127e5cc
Adding ALGEBRAIC strategy to `test_curve_curve`.
Also deleting `test_implicitized_intersect`.
