+++ functional_tests/test_curve_curve.py
@@ -19,8 +19,9 @@ import numpy as np
 import pytest
 import six
 
-import bezier.curve
+from bezier import _implicitization
 from bezier import _intersection_helpers
+import bezier.curve
 
 import runtime_utils
 from runtime_utils import CurveIntersectionType
@@ -29,14 +30,28 @@ from runtime_utils import CurveIntersectionType
 SPACING = np.spacing  # pylint: disable=no-member
 CONFIG = runtime_utils.Config()
 GEOMETRIC = bezier.curve.IntersectionStrategy.geometric
+ALGEBRAIC = bezier.curve.IntersectionStrategy.algebraic
 S_PROP = operator.attrgetter('s')
 _, INTERSECTIONS = runtime_utils.curve_intersections_info()
+INCORRECT_VALUES_MSG = """\
+Multipliers were:
+{}
+
+However, the actual observed (ULP relative) errors were:
+{}
+
+The issue seems to be in the following index pair(s):
+{}
+"""
 ULPS_ALLOWED = 3.0
 # NOTE: Set a threshold for values that are "approximately" zero.
 #       This is an **absolute** error rather than a relative error
 #       since relative to zero error is always infinite.
 ZERO_THRESHOLD = 0.5**52
 ZERO_MISS = object()
+# NOTE: We use units of least precision (ULP) as error. These
+#       are for the very rare cases where the computed values
+#       differ from the actual values by more than 3 ULPs.
 ULPS_ALLOWED_OVERRIDE = {
     GEOMETRIC: {
         12: {
@@ -83,6 +98,44 @@ ULPS_ALLOWED_OVERRIDE = {
             (0, 1): 1013,  # Established on Ubuntu 16.04
         },
     },
+    ALGEBRAIC: {
+        2: {
+            (1, 3): 4,  # Established on Ubuntu 16.04
+        },
+        17: {
+            (0, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+        },
+        18: {
+            (0, 6): 5,  # Established on Ubuntu 16.04
+            (1, 6): 4,  # Established on Ubuntu 16.04
+            (3, 3): ZERO_MISS,  # Established on Ubuntu 16.04
+            (3, 5): ZERO_MISS,  # Established on Ubuntu 16.04
+            (3, 7): ZERO_MISS,  # Established on Ubuntu 16.04
+        },
+        22: {
+            (0, 0): 12,  # Established on Ubuntu 16.04
+            (0, 1): 29,  # Established on Ubuntu 16.04
+            (0, 2): 9,  # Established on Ubuntu 16.04
+        },
+        23: {
+            (0, 1): 4,  # Established on Ubuntu 16.04
+        },
+        25: {
+            (0, 6): ZERO_MISS,  # Established on Ubuntu 16.04
+        },
+        37: {
+            (0, 1): 165,  # Established on Ubuntu 16.04
+        },
+        38: {
+            (0, 1): 18,  # Established on Ubuntu 16.04
+        },
+        39: {
+            (0, 1): 165,  # Established on Ubuntu 16.04
+        },
+        40: {
+            (0, 1): 18,  # Established on Ubuntu 16.04
+        },
+    },
 }
 TANGENT_SUCCESS = {
     GEOMETRIC: (
@@ -91,6 +144,7 @@ TANGENT_SUCCESS = {
         19,
         41,
     ),
+    ALGEBRAIC: ()
 }
 COINCIDENT_SUCCESS = {
     GEOMETRIC: (
@@ -98,11 +152,13 @@ COINCIDENT_SUCCESS = {
         34,
         35,
     ),
+    ALGEBRAIC: ()
 }
 INCORRECT_COUNT = {
     GEOMETRIC: (
[…]
