### 2017-07-25 a13269c
Adding unit test for recently added Intersection.__dict__.
