### 2017-07-25 8927607
Adding "stricter" exception checking in fnl. tests.
