### 2017-07-25 513b318
Removing unused `CONFIG` in `test_curve_curve`.
