### 2017-07-26 #267 @tacaswell
All of my issues have been addressed.

I recommend this to be published!

Sorry for being so slow on my end.
