### 2017-07-26 #267 @tacaswell
There are far more references in the bibliography than cited from paper.md, does that get sorted out at somepoint in the next step?
