### 2017-07-26 #267 @arfon
Thanks @tacaswell.

> There are far more references in the bibliography than cited from paper.md, does that get sorted out at somepoint in the next step?

@dhermes - all of the references need moving to the `paper.bib` file and citing directly. (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))

In addition, could you please expand a little on the `paper.md` file? We (the JOSS editors) are in the process of adding more guidance for authors, including a better indication of expected paper length. Currently it's hard to get a picture of the functionality of the software, and also include a statement of need, in less than 250 words. Would you be able to put a little more detail in your submission please?
