### 2017-07-27 1cbdfba
Make sure get_random_nodes() returns Fortran contiguous.
