### 2017-07-27 b9a7a8c
Using cross-product speedup when possible.
Also dropping ``degree`` from ``linearization_error()``.
