+++ tests/test__intersection_helpers.py
@@ -143,17 +143,17 @@ class Test_speedup_bbox_intersect(Test__bbox_intersect):
 class Test__linearization_error(unittest.TestCase):
 
     @staticmethod
-    def _call_function_under_test(nodes, degree):
+    def _call_function_under_test(nodes):
         from bezier import _intersection_helpers
 
-        return _intersection_helpers._linearization_error(nodes, degree)
+        return _intersection_helpers._linearization_error(nodes)
 
     def test_linear(self):
         nodes = np.asfortranarray([
             [0.0, 0.0],
             [1.0, 2.0],
         ])
-        error_val = self._call_function_under_test(nodes, 1)
+        error_val = self._call_function_under_test(nodes)
         self.assertEqual(error_val, 0.0)
 
     def test_degree_elevated_linear(self):
@@ -162,7 +162,7 @@ class Test__linearization_error(unittest.TestCase):
             [0.5, 1.0],
             [1.0, 2.0],
         ])
-        error_val = self._call_function_under_test(nodes, 2)
+        error_val = self._call_function_under_test(nodes)
         self.assertEqual(error_val, 0.0)
 
         nodes = np.asfortranarray([
@@ -172,7 +172,7 @@ class Test__linearization_error(unittest.TestCase):
             [0.75, 1.5],
             [1.0, 2.0],
         ])
-        error_val = self._call_function_under_test(nodes, 4)
+        error_val = self._call_function_under_test(nodes)
         self.assertEqual(error_val, 0.0)
 
     def test_hidden_linear(self):
@@ -183,7 +183,7 @@ class Test__linearization_error(unittest.TestCase):
             [6.0, 8.0],
             [3.0, 4.0],
         ])
-        error_val = self._call_function_under_test(nodes, 2)
+        error_val = self._call_function_under_test(nodes)
         # D^2 v = [-9, -12]
         expected = 0.125 * 2 * 1 * 15.0
         self.assertEqual(error_val, expected)
@@ -200,7 +200,7 @@ class Test__linearization_error(unittest.TestCase):
         #             d Nodes = [1, 1], [4, 5]
         #           d^2 Nodes = [3, 4]
         #       so that sqrt(3^2 + 4^2) = 5.0
-        error_val = self._call_function_under_test(nodes, 2)
+        error_val = self._call_function_under_test(nodes)
         expected = 0.125 * 2 * 1 * 5.0
         self.assertEqual(error_val, expected)
 
@@ -209,8 +209,8 @@ class Test__linearization_error(unittest.TestCase):
         # of (1/2)^2 = 4.
         curve = bezier.Curve(nodes, degree=2)
         left, right = curve.subdivide()
-        error_left = self._call_function_under_test(left._nodes, 2)
-        error_right = self._call_function_under_test(right._nodes, 2)
+        error_left = self._call_function_under_test(left._nodes)
+        error_right = self._call_function_under_test(right._nodes)
         self.assertEqual(error_left, 0.25 * expected)
         self.assertEqual(error_right, 0.25 * expected)
 
@@ -224,7 +224,7 @@ class Test__linearization_error(unittest.TestCase):
         #             d Nodes = [2, -5, 4], [5, 7, 0]
         #           d^2 Nodes = [3, 12, -4]
         #       so that sqrt(3^2 + 12^2 + 4^2) = 13.0
-        error_val = self._call_function_under_test(nodes, 2)
+        error_val = self._call_function_under_test(nodes)
         expected = 0.125 * 2 * 1 * 13.0
         self.assertEqual(error_val, expected)
 
[…]
