### 2017-07-27 #267 @dhermes
> could you please expand a little on the `paper.md` file

Will do!

Should I mention features the library **doesn't** have (that could be added at a later date)? Following a standard text / course syllabus I could outline the "toolbox" one would typically want when dealing with Bezier curves and how it compares to the feature set of the library.

----

@arfon IIUC, when @tacaswell said

> There are far more references in the bibliography than cited from `paper.md`

he meant that `paper.bib` has more references (23):

```
$ cat paper.bib | grep -e '^@'
@article{Farouki1987,
@article{Sederberg1989,
@article{Sederberg1989-alg,
@article{Sederberg1990,
@article{Farouki1991,
@article{Dunavant1985,
@article{Farouki1988,
@article{Farouki1990,
@article{Farouki1996,
@article{Kim1998,
@article{Sederberg1986,
@inproceedings{Roth,
@misc{Roth2000,
@article{Farin1986,
@article{Sederberg1986-improp,
@article{Sederberg1984,
@article{Brown1971,
@article{Farouki2012,
@article{Bus2008,
@article{JonssonVavasis,
@techreport{Manocha:CSD-92-698,
@book{Farin2001,
@misc{SederbergNotes,
$ cat paper.bib | grep -e '^@' | wc -l
23
```

than are actually used in `paper.md`:

```
# References

- [Lecture notes][3]: Computer Aided Geometric Design by Thomas W.
  Sederberg of BYU
- [Curves and Surfaces for CAGD][4] by Gerald Farin

...

[3]: http://tom.cs.byu.edu/~557/text/cagd.pdf
[4]: https://www.amazon.com/Curves-Surfaces-CAGD-Fifth-Practical/dp/1558607374
```
