### 2017-07-27 8e8b647
Updating classifiers in `setup.py`.
