### 2017-07-27 75164f0
Updating tags in `paper.md`.
The tags were previously just copy-pasted from the example (shame
on me).
