### 2017-07-27 fdf986b
Adding citation (to `paper.bib` references) to `paper.md`.
H/T: http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citations
