### 2017-07-27 2a9aa22
Making sure Linearization.(start|end)_node are Fortran contiguous.
This ends up being a non-trivial performance hit because of the
new array creation:

    In [1]: import numpy as np

    In [2]: values = np.asfortranarray([
       ...:     [0.5, 1.0],
       ...:     [10.0, 1.0],
       ...:     [-1.0, 4.25],
       ...:     [0.0, -10.75],
       ...: ])

    In [3]: def f(nodes):
       ...:     return nodes[0, :].reshape((1, 2), order='F')
       ...:

    In [4]: def g(nodes):
       ...:     return np.asfortranarray([
       ...:         [nodes[0, 0], nodes[0, 1]],
       ...:     ])
    
[…]
