### 2017-07-27 a91b54a
Making sure array is Fortran contiguous before passing to `evaluate_multi_barycentric()`.
