### 2017-07-27 55070d2
Renaming `Test_speedup_newton_refine`.
Changed to `Test_speedup_newton_refine_intersect` to more closely
describe the name of the function being tested.
