### 2017-07-27 d9a5f49
Updating unit tests for (start|end)_node.
Was broken by a91b54a2d7e77f2b5e367624830ac8bf8b55b363.
