### 2017-07-27 c63973e
Making sure to use Fortran contiguous random arrays in unit test.
