### 2017-07-28 d9e3b59
Lint fix: allow more lines in source.
