### 2017-07-29 #267 @arfon
> Should I mention features the library doesn't have (that could be added at a later date)? Following a standard text / course syllabus I could outline the "toolbox" one would typically want when dealing with Bezier curves and how it compares to the feature set of the library.

I think the paper should only describe the current functionality at a high level rather than write about what the library _might_ do one day.

Also, please note, for the citations, the section after the `# References` heading should be completely empty. Pandoc compiles the references you have cited in your `paper.md` from the `paper.bib` file into the final output. You can use [this paper](https://github.com/arfon/fidgit/blob/master/paper/paper.md) as an example.
