### 2017-08-01 090e970
General docs / config maintenance.
Most of the changes occurred when I was reviewing the rendered
docs for Surface and Curve (while editing `paper.md`).

Among the changes:

- Setting a `dev` release in `setup.py`
- Updating `surface_subdivide2.png` to include all 6 control points
  rather than just the 5 that determine the control polygon
- Updating `make_surface_surface_images.py` to use the newer
  attribute name from `SurfaceIntersectionsInfo`
- Using `surface(1|2)` aliases in `test_surface_surface.py`
- Adding spaces in "to / from" (vs. "to/from")
- Adding a note about how our data layout
[…]
