### 2017-08-01 e6c4536
Removing hard-coded references in `paper.md`.
Done per: https://github.com/openjournals/joss-reviews/issues/267#issuecomment-318829704
