### 2017-08-01 975ac6b
Re-write of summary in `paper.md`.
This is intended to indicate:

- what Bezier curves / surfaces are
- what they are used for / who uses them
- what algorithms are supported
