+++ paper.bib
@@ -13,19 +13,6 @@
   journal = {Computer Aided Geometric Design}
 }
 @article{Sederberg1989,
-  doi = {10.1016/0010-4485(89)90058-4},
-  url = {https://dx.doi.org/10.1016/0010-4485(89)90058-4},
-  year  = {1989},
-  month = {Oct},
-  publisher = {Elsevier {BV}},
-  volume = {21},
-  number = {8},
-  pages = {505--508},
-  author = {T.W. Sederberg and H.N. Christiansen and S. Katz},
-  title = {Improved test for closed loops in surface intersections},
-  journal = {Computer-Aided Design}
-}
-@article{Sederberg1989-alg,
   doi = {10.1016/0010-4485(89)90015-8},
   url = {https://dx.doi.org/10.1016/0010-4485(89)90015-8},
   year  = {1989},
@@ -64,35 +51,9 @@
   title = {On the stability of transformations between power and {B}ernstein polynomial forms},
   journal = {Computer Aided Geometric Design}
 }
-@article{Dunavant1985,
-  doi = {10.1002/nme.1620210612},
-  url = {https://dx.doi.org/10.1002/nme.1620210612},
-  year  = {1985},
-  month = {Jun},
-  publisher = {Wiley-Blackwell},
-  volume = {21},
-  number = {6},
-  pages = {1129--1148},
-  author = {D. A. Dunavant},
-  title = {High degree efficient symmetrical {G}aussian quadrature rules for the triangle},
-  journal = {International Journal for Numerical Methods in Engineering}
-}
-@article{Farouki1988,
-  doi = {10.1016/0167-8396(88)90016-7},
-  url = {https://dx.doi.org/10.1016/0167-8396(88)90016-7},
-  year  = {1988},
-  month = {Jun},
-  publisher = {Elsevier {BV}},
-  volume = {5},
-  number = {1},
-  pages = {1--26},
-  author = {R.T. Farouki and V.T. Rajan},
-  title = {Algorithms for polynomials in {B}ernstein form},
-  journal = {Computer Aided Geometric Design}
-}
 @article{Farouki1990,
   doi = {10.2307/2008438},
-  url = {https://doi.org/10.2307%2F2008438},
+  url = {https://doi.org/10.2307/2008438},
   year  = {1990},
   month = {oct},
   publisher = {{JSTOR}},
@@ -105,7 +66,7 @@
 }
 @article{Farouki1996,
   doi = {10.1090/s0025-5718-96-00759-4},
-  url = {https://doi.org/10.1090%2Fs0025-5718-96-00759-4},
+  url = {https://doi.org/10.1090/s0025-5718-96-00759-4},
   year  = {1996},
   month = {oct},
   publisher = {American Mathematical Society ({AMS})},
@@ -142,75 +103,6 @@
   title = {Comparison of three curve intersection algorithms},
   journal = {Computer-Aided Design}
 }
-@inproceedings{Roth,
-  doi = {10.1109/pccga.2000.883971},
-  url = {https://dx.doi.org/10.1109/PCCGA.2000.883971},
-  publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
-  author = {S.H.M. Roth and P. Diezi and M.H. Gross},
-  title = {Triangular {B}{\'{e}}zier clipping},
-  booktitle = {Proceedings the Eighth Pacific Conference on Computer Graphics and Applications},
-  year = {2000}
-}
-@misc{Roth2000,
-  doi = {10.3929/ethz-a-006654115},
-  url = {https://doi.org/10.3929/ethz-a-006654115},
[…]
