### 2017-08-01 #267 @dhermes
I [just updated][1] the summary. LMK what you think @arfon and @tacaswell?

The 3 paragraphs are:

- What are B&#xe9;zier curves / surfaces
- Where (i.e. what fields) have use for them? (This is essentially equivalent to stating the "audience")
- What features are present for people who might use them

[1]: https://github.com/dhermes/bezier/commit/975ac6b1a4313db4dcdc17396d6d34561005939e
