### 2017-08-02 26cf2c1
Adding in_interval() to Fortan speedup.
Also changing branching in `intersect_one_round()` to be
easier to read / understand.

Also making slight update to `Makefile` target for the
auto-generated pyf file and then re-running it.
