+++ src/bezier/speedup.f90
@@ -21,11 +21,14 @@ module speedup
        specialize_curve_generic, specialize_curve_quadratic, &
        specialize_curve, jacobian_both, evaluate_hodograph, &
        newton_refine_intersect, jacobian_det, bbox_intersect, &
-       wiggle_interval, parallel_different, from_linearized
+       wiggle_interval, parallel_different, from_linearized, bbox_line_intersect
 
   ! NOTE: This still relies on .f2py_f2cmap being present
   !       in the directory that build is called from.
   integer, parameter :: dp=kind(0.d0)
+  integer, parameter :: BoxIntersectionType_INTERSECTION = 0
+  integer, parameter :: BoxIntersectionType_TANGENT = 1
+  integer, parameter :: BoxIntersectionType_DISJOINT = 2
 
 contains
 
@@ -652,13 +655,13 @@ contains
     if ( &
          right2 < left1 .OR. right1 < left2 .OR. &
          top2 < bottom1 .OR. top1 < bottom2) then
-        enum_ = 2
-     else if ( &
-          right2 == left1 .OR. right1 == left2 .OR. &
-          top2 == bottom1 .OR. top1 == bottom2) then
-        enum_ = 1
+       enum_ = BoxIntersectionType_DISJOINT
+    else if ( &
+         right2 == left1 .OR. right1 == left2 .OR. &
+         top2 == bottom1 .OR. top1 == bottom2) then
+       enum_ = BoxIntersectionType_TANGENT
     else
-        enum_ = 0
+       enum_ = BoxIntersectionType_INTERSECTION
     end if
 
   end subroutine bbox_intersect
@@ -796,7 +799,7 @@ contains
        else
           call bbox_intersect( &
                degree1 + 1, nodes1, degree2 + 1, nodes2, enum_)
-          if (enum_ == 2) then  ! Disjoint
+          if (enum_ == BoxIntersectionType_DISJOINT) then
              does_intersect = .FALSE.
              return
           end if
@@ -835,7 +838,7 @@ contains
 
   end subroutine from_linearized
 
-  subroutine in_interval(value_, start, end, result_)
+  pure subroutine in_interval(value_, start, end, result_)
 
     real(dp), intent(in) :: value_, start, end
     logical(1), intent(out) :: result_
@@ -843,5 +846,104 @@ contains
     result_ = (start <= value_) .AND. (value_ <= end)
 
   end subroutine in_interval
+
+  subroutine bbox_line_intersect( &
+       num_nodes, nodes, line_start, line_end, enum_)
+
+    !f2py integer intent(hide), depend(nodes) :: num_nodes = size(nodes, 1)
+    integer :: num_nodes
+    real(dp), intent(in) :: nodes(num_nodes, 2)
+    real(dp), intent(in) :: line_start(1, 2)
+    real(dp), intent(in) :: line_end(1, 2)
+    integer, intent(out) :: enum_
+    ! Variables outside of signature.
+    real(dp) :: left, right, bottom, top
+    real(dp) :: segment_start(1, 2)
+    real(dp) :: segment_end(1, 2)
+    real(dp) :: s_curr, t_curr
+    logical(1) :: success, predicate1, predicate2
+
+    call bbox(num_nodes, nodes, left, right, bottom, top)
+
+    ! Check if line start is inside the bounding box.
+    call in_interval(line_start(1, 1), left, right, predicate1)
+    call in_interval(line_start(1, 2), bottom, top, predicate2)
+    if (predicate1 .AND. predicate2) then
+       enum_ = BoxIntersectionType_INTERSECTION
+       return
+    end if
+
+    ! Check if line end is inside the bounding box.
+    call in_interval(line_end(1, 1), left, right, predicate1)
+    call in_interval(line_end(1, 2), bottom, top, predicate2)
[…]
