+++ tests/test__intersection_helpers.py
@@ -10,15 +10,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import itertools
 import unittest
 
 import mock
 import numpy as np
+import six
 
 from tests import utils
 
 
 SPACING = np.spacing  # pylint: disable=no-member
+UNIT_SQUARE = np.asfortranarray([
+    [0.0, 0.0],
+    [1.0, 0.0],
+    [1.0, 1.0],
+    [0.0, 1.0],
+])
 
 
 def check_intersection(test_case, intersection, expected,
@@ -75,13 +83,6 @@ class Test__check_close(utils.NumPyTestCase):
 
 class Test__bbox_intersect(unittest.TestCase):
 
-    UNIT_SQUARE = np.asfortranarray([
-        [0.0, 0.0],
-        [1.0, 0.0],
-        [1.0, 1.0],
-        [0.0, 1.0],
-    ])
-
     @staticmethod
     def _call_function_under_test(nodes1, nodes2):
         from bezier import _intersection_helpers
@@ -91,32 +92,32 @@ class Test__bbox_intersect(unittest.TestCase):
     def test_intersect(self):
         from bezier import _intersection_helpers
 
-        nodes = self.UNIT_SQUARE + np.asfortranarray([[0.5, 0.5]])
-        result = self._call_function_under_test(self.UNIT_SQUARE, nodes)
+        nodes = UNIT_SQUARE + np.asfortranarray([[0.5, 0.5]])
+        result = self._call_function_under_test(UNIT_SQUARE, nodes)
         expected = _intersection_helpers.BoxIntersectionType.INTERSECTION
         self.assertEqual(result, expected)
 
     def test_far_apart(self):
         from bezier import _intersection_helpers
 
-        nodes = self.UNIT_SQUARE + np.asfortranarray([[100.0, 100.0]])
-        result = self._call_function_under_test(self.UNIT_SQUARE, nodes)
+        nodes = UNIT_SQUARE + np.asfortranarray([[100.0, 100.0]])
+        result = self._call_function_under_test(UNIT_SQUARE, nodes)
         expected = _intersection_helpers.BoxIntersectionType.DISJOINT
         self.assertEqual(result, expected)
 
     def test_disjoint_but_aligned(self):
         from bezier import _intersection_helpers
 
-        nodes = self.UNIT_SQUARE + np.asfortranarray([[1.0, 2.0]])
-        result = self._call_function_under_test(self.UNIT_SQUARE, nodes)
+        nodes = UNIT_SQUARE + np.asfortranarray([[1.0, 2.0]])
+        result = self._call_function_under_test(UNIT_SQUARE, nodes)
         expected = _intersection_helpers.BoxIntersectionType.DISJOINT
         self.assertEqual(result, expected)
 
     def test_tangent(self):
         from bezier import _intersection_helpers
 
-        nodes = self.UNIT_SQUARE + np.asfortranarray([[1.0, 0.0]])
-        result = self._call_function_under_test(self.UNIT_SQUARE, nodes)
+        nodes = UNIT_SQUARE + np.asfortranarray([[1.0, 0.0]])
+        result = self._call_function_under_test(UNIT_SQUARE, nodes)
         expected = _intersection_helpers.BoxIntersectionType.TANGENT
         self.assertEqual(result, expected)
 
@@ -124,8 +125,8 @@ class Test__bbox_intersect(unittest.TestCase):
         from bezier import _intersection_helpers
 
         x_val = 1.0 + SPACING(1.0)
-        nodes = self.UNIT_SQUARE + np.asfortranarray([[x_val, 0.0]])
[…]
