### 2017-08-02 85fe16c
Converting Fortran in_interval() from subroutine to function.
In general, functions are too difficult to wrap with `f2py`
but we don't wrap `is_interval()` so it's easier to make
it a function (easier so we don't have to call it and dump
into a dummy variable).
