### 2017-08-02 b89823e
Adding Zenodo badge to project.
I bounced back and forth between using the explicit link:

    https://zenodo.org/record/838307

or the "recommended" link.

    https://zenodo.org/badge/latestdoi/73047402

I was also unsure if I should use the "dynamic" badge that
redirects to latest (currently ``10.5281/zenodo.838307.svg``):

    https://zenodo.org/badge/73047402.svg

or the "all of the versions" DOI badge the

    https://zenodo.org/badge/DOI/10.5281/zenodo.838307.svg

But @will-e-p convinced me that using the "dynamic" links
was the "best" choice.
