### 2017-08-02 #267 @dhermes
Sure thing: [`10.5281/zenodo.838308`][1] is for `bezier==0.4.0` ([`10.5281/zenodo.838307`][2] is the "catch-all" DOI for the project)

[1]: https://zenodo.org/record/838308
[2]: https://zenodo.org/record/838307
