### 2017-08-02 c698929
Updating README template for Zenodo badge.
