### 2017-08-02 de5f32f
Adding "diff" to error message in `check_readme`.
