### 2017-08-02 #267 @arfon
@tacaswell many thanks for your review here ✨

@dhermes - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00267 ⚡️ 🚀 :boom:
